### 2018-06-04 #760 @whedon
[PRE REVIEW]: quanteda: An R package for the quantitative analysis of textual data
**Submitting author:** @kbenoit (<a href="http://orcid.org/0000-0002-0797-564X">Kenneth Benoit</a>)
**Repository:** <a href="https://github.com/quanteda/quanteda" target ="_blank">https://github.com/quanteda/quanteda</a>
**Version:** v1.3
