### 2018-06-04 #760 @arfon
:wave: @kenbenoit - can you take a look the list of JOSS reviewers https://bit.ly/joss-reviewers and suggest a few potential reviewers?
