### 2018-06-05 b449f79
Fix link to quick start
