+++ README.Rmd
@@ -25,7 +25,7 @@ knitr::opts_chunk$set(
 An R package for managing and analyzing text, created by [Kenneth Benoit](http://kenbenoit.net).
  Supported by the European Research Council grant ERC-2011-StG 283794-QUANTESS.
 
-For more details, see http://docs.quanteda.io/index.html.
+For more details, see https://docs.quanteda.io/index.html.
 
 ## How to Install
 
