### 2018-06-05 c07e230
Add tidytext to Suggests
