+++ DESCRIPTION
@@ -32,49 +32,50 @@ Depends:
     R (>= 3.1.0),
     methods
 Imports:
-    extrafont,
-    Matrix (>= 1.2),
     data.table (>= 1.9.6),
-    SnowballC,
-    sna,
-    ggrepel,
-    network,
-    Rcpp (>= 0.12.12),
-    RcppParallel,
-    RSpectra,
-    stringi,
+    extrafont,
     fastmatch,
     ggplot2 (>= 2.2.0),
-    xml2,
-    yaml,
+    ggrepel,
     lubridate,
     magrittr,
+    Matrix (>= 1.2),
+    network,
+    RSpectra,
+    Rcpp (>= 0.12.12),
+    RcppParallel,
+    sna,
+    SnowballC,
     spacyr,
-    stopwords
+    stopwords,
+    stringi,
+    xml2,
+    yaml
 LinkingTo: Rcpp, RcppParallel, RcppArmadillo (>= 0.7.600.1.0)
 Suggests:
-    knitr,
-    rmarkdown,
-    lda,
-    proxy,
-    topicmodels,
-    tm (>= 0.6),
-    slam,
-    testthat,
-    RColorBrewer,
-    xtable,
-    DT,
     ca,
-    purrr,
-    ExPosition,     
     dplyr,     
+    DT,
     e1071,     
+    ExPosition,     
     koRpus,     
+    lda,
     lsa,     
+    proxy,
+    purrr,
+    RColorBrewer,
+    rmarkdown,
+    slam,
     stm,     
     svs,     
-    tibble,     
-    text2vec
+    testthat,
+    text2vec,
+    tibble,
+    tidytext,
+    tm (>= 0.6),
+    topicmodels,
+    xtable,
+    knitr
 URL: http://quanteda.io
 Encoding: UTF-8
 BugReports: https://github.com/quanteda/quanteda/issues
