### 2018-06-07 624d982
Add as.igraph.fcm() - It works in the same way as as.network for #1260 - Both as.network() and as.igraph() return marginal feqture frequency as vertex attribute
