+++ NAMESPACE
@@ -51,6 +51,7 @@ S3method(as.dfm,textmodel_lsa)
 S3method(as.dictionary,data.frame)
 S3method(as.dictionary,default)
 S3method(as.dist,dist)
+S3method(as.igraph,fcm)
 S3method(as.list,dist)
 S3method(as.list,dist_selection)
 S3method(as.list,tokens)
@@ -386,6 +387,7 @@ export(as.coefficients_textmodel)
 export(as.corpus)
 export(as.dfm)
 export(as.dictionary)
+export(as.igraph)
 export(as.network)
 export(as.statistics_textmodel)
 export(as.summary.textmodel)
