### 2018-06-07 0d81898
Add tests for as.igraph.fcm() and as.network.fcm()
