+++ man/as.igraph.Rd
@@ -1,0 +1,17 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/textplot_network.R
+\name{as.igraph}
+\alias{as.igraph}
+\title{redefinition of igraph::as.igraph()}
+\usage{
+as.igraph(x, ...)
+}
+\arguments{
+\item{x}{input object}
+
+\item{...}{additional arguments}
+}
+\description{
+redefinition of igraph::as.igraph()
+}
+\keyword{internal}
