### 2018-06-07 88175d0
Separate plot_network
