+++ R/textplot_network.R
@@ -18,7 +18,8 @@
 #' @param offset if \code{NULL}, the distance between vertices and texts are
 #'   determined automatically.
 #' @param vertex_labelfont font-family of texts. Use default font if \code{NULL}.
-#' @param ... additional arguments passed to \link[network]{network}.
+#' @param ... additional arguments passed to \link[network]{network} or
+#'   \link[igraph]{graph_from_adjacency_matrix}.
 #' @details Currently the size of the network is limited to 1000, because of the
 #'   computationally intensive nature of network formation for larger matrices.
 #'   When the \link{fcm} is large, users should select features using
@@ -45,8 +46,8 @@ textplot_network <- function(x, min_freq = 0.5, omit_isolated = TRUE,
                              edge_size = 2, 
                              vertex_color = "#4D4D4D", vertex_size = 2,
                              vertex_labelcolor = NULL,
-                             offset = NULL, 
-                             vertex_labelfont = NULL, ...) {
+                             vertex_labelfont = NULL, 
+                             offset = NULL, ...) {
     UseMethod("textplot_network")
 }
 
@@ -56,16 +57,16 @@ textplot_network.dfm <- function(x, min_freq = 0.5, omit_isolated = TRUE,
                                  edge_size = 2, 
                                  vertex_color = "#4D4D4D", vertex_size = 2,
                                  vertex_labelcolor = NULL,
-                                 offset = NULL, 
-                                 vertex_labelfont = NULL, ...) {
+                                 vertex_labelfont = NULL, 
+                                 offset = NULL, ...) {
 
     textplot_network(fcm(x), min_freq = min_freq, omit_isolated = omit_isolated, 
                      edge_color = edge_color, edge_alpha = edge_alpha, 
                      edge_size = edge_size, 
                      vertex_color = vertex_color, vertex_size = vertex_size,
                      vertex_labelcolor = vertex_labelcolor,
-                     offset = NULL, 
-                     vertex_labelfont = vertex_labelfont, ...)
+                     vertex_labelfont = vertex_labelfont, 
+                     offset = NULL, ...)
 }
 
     
@@ -75,22 +76,21 @@ textplot_network.fcm <- function(x, min_freq = 0.5, omit_isolated = TRUE,
                                  edge_size = 2, 
                                  vertex_color = "#4D4D4D", vertex_size = 2,
                                  vertex_labelcolor = NULL,
+                                 vertex_labelfont = NULL, 
                                  offset = NULL, 
-                                 vertex_labelfont = NULL, ...) {
+                                 ...) {
     
-    label <- x1 <- x2 <- y <- y1 <- y2 <- NULL
-    
-    vertex_labelfont <- check_font(vertex_labelfont)
-    n <- as.network(x, min_freq = min_freq, omit_isolated = omit_isolated, ...)
+    font <- check_font(vertex_labelfont)
+    net <- as.network(x, min_freq = min_freq, omit_isolated = omit_isolated, ...)
 
-    vertex <- data.frame(sna::gplot.layout.fruchtermanreingold(n, NULL))
+    vertex <- data.frame(sna::gplot.layout.fruchtermanreingold(net, NULL))
     colnames(vertex) <- c("x", "y")
-    vertex$label <- network::network.vertex.names(n)
+    vertex$label <- network::network.vertex.names(net)
 
-    weight <- network::get.edge.attribute(n, "weight")
+    weight <- network::get.edge.attribute(net, "weight")
     weight <- weight / max(weight)
[…]
