+++ man/textplot_network.Rd
@@ -9,7 +9,7 @@
 textplot_network(x, min_freq = 0.5, omit_isolated = TRUE,
   edge_color = "#1F78B4", edge_alpha = 0.5, edge_size = 2,
   vertex_color = "#4D4D4D", vertex_size = 2, vertex_labelcolor = NULL,
-  offset = NULL, vertex_labelfont = NULL, ...)
+  vertex_labelfont = NULL, offset = NULL, ...)
 
 \method{as.network}{fcm}(x, min_freq = 0.5, omit_isolated = TRUE, ...)
 
@@ -38,12 +38,13 @@ other edges are determined proportionally to the highest frequency.}
 \item{vertex_labelcolor}{color of texts. Defaults to the same as
 \code{vertex_color}. If \code{NA} is given, texts are not rendered.}
 
+\item{vertex_labelfont}{font-family of texts. Use default font if \code{NULL}.}
+
 \item{offset}{if \code{NULL}, the distance between vertices and texts are
 determined automatically.}
 
-\item{vertex_labelfont}{font-family of texts. Use default font if \code{NULL}.}
-
-\item{...}{additional arguments passed to \link[network]{network}.}
+\item{...}{additional arguments passed to \link[network]{network} or
+\link[igraph]{graph_from_adjacency_matrix}.}
 }
 \description{
 Plot an \link{fcm} object as a network, where edges show co-occurrences of
@@ -70,6 +71,10 @@ fcm_select(myfcm, feat, verbose = FALSE) \%>\%
 }
 \seealso{
 \code{\link{fcm}}
+
+\code{\link[network]{network}}
+
+\code{\link[igraph]{graph_from_adjacency_matrix}}
 }
 \author{
 Kohei Watanabe and Stefan Müller
