### 2018-06-08 #760 @lmullen
I can review this but may need a little flexibility with the due date.
