### 2018-06-11 #760 @arfon
@borishejblum - would you be willing to review this submission for JOSS?
