+++ README.md
@@ -1,17 +1,29 @@
 
-[![quanteda: quantitative analysis of textual data](https://cdn.rawgit.com/quanteda/quanteda/master/images/quanteda_logo.svg)](http://quanteda.io)
+[![quanteda: quantitative analysis of textual
+data](https://cdn.rawgit.com/quanteda/quanteda/master/images/quanteda_logo.svg)](http://quanteda.io)
 
-[![CRAN Version](https://www.r-pkg.org/badges/version/quanteda)](https://CRAN.R-project.org/package=quanteda) [![Downloads](https://cranlogs.r-pkg.org/badges/quanteda)](https://CRAN.R-project.org/package=quanteda) [![Total Downloads](https://cranlogs.r-pkg.org/badges/grand-total/quanteda?color=orange)](https://CRAN.R-project.org/package=quanteda) [![Travis-CI Build Status](https://travis-ci.org/quanteda/quanteda.svg?branch=master)](https://travis-ci.org/quanteda/quanteda) [![Appveyor Build status](https://ci.appveyor.com/api/projects/status/e3tf2h1ff0nlv249/branch/master?svg=true)](https://ci.appveyor.com/project/kbenoit/quanteda/branch/master) [![codecov](https://codecov.io/gh/quanteda/quanteda/branch/master/graph/badge.svg)](https://codecov.io/gh/quanteda/quanteda) [![DOI](https://zenodo.org/badge/5424649.svg)](https://zenodo.org/badge/latestdoi/5424649)
+[![CRAN
+Version](https://www.r-pkg.org/badges/version/quanteda)](https://CRAN.R-project.org/package=quanteda)
+[![Downloads](https://cranlogs.r-pkg.org/badges/quanteda)](https://CRAN.R-project.org/package=quanteda)
+[![Total
+Downloads](https://cranlogs.r-pkg.org/badges/grand-total/quanteda?color=orange)](https://CRAN.R-project.org/package=quanteda)
+[![Travis-CI Build
+Status](https://travis-ci.org/quanteda/quanteda.svg?branch=master)](https://travis-ci.org/quanteda/quanteda)
+[![Appveyor Build
+status](https://ci.appveyor.com/api/projects/status/e3tf2h1ff0nlv249/branch/master?svg=true)](https://ci.appveyor.com/project/kbenoit/quanteda/branch/master)
+[![codecov](https://codecov.io/gh/quanteda/quanteda/branch/master/graph/badge.svg)](https://codecov.io/gh/quanteda/quanteda)
+[![DOI](https://zenodo.org/badge/5424649.svg)](https://zenodo.org/badge/latestdoi/5424649)
+[![status](http://joss.theoj.org/papers/40b988ba4827f8fdc07a29351c2f74b8/status.svg)](http://joss.theoj.org/papers/40b988ba4827f8fdc07a29351c2f74b8)
 
-About
------
+## About
 
-An R package for managing and analyzing text, created by [Kenneth Benoit](http://kenbenoit.net). Supported by the European Research Council grant ERC-2011-StG 283794-QUANTESS.
+An R package for managing and analyzing text, created by [Kenneth
+Benoit](http://kenbenoit.net). Supported by the European Research
+Council grant ERC-2011-StG 283794-QUANTESS.
 
 For more details, see <https://docs.quanteda.io/index.html>.
 
-How to Install
---------------
+## How to Install
 
 The normal way from CRAN, using your R GUI or
 
@@ -26,28 +38,44 @@ Or for the latest development version:
 devtools::install_github("quanteda/quanteda") 
 ```
 
-Because this compiles some C++ and Fortran source code, you will need to have installed the appropriate compilers.
+Because this compiles some C++ and Fortran source code, you will need to
+have installed the appropriate compilers.
 
[…]
