+++ paper/paper.md
@@ -13,7 +13,8 @@ authors:
 - affiliation: 3
   name: Paul Nulty
   orcid: 0000-0002-7214-4666
-- name: Adam Obeng
+- affiliation: 5
+  name: Adam Obeng
   orcid: 0000-0002-2906-4775
 - affiliation: 4
   name: Stefan Müller
@@ -35,6 +36,8 @@ affiliations:
   name: Centre for Research in Arts, Social Science and Humanities, University of Cambridge
 - index: 4
   name: Department of Political Science, Trinity College Dublin
+- index: 5
+  name: Facebook, Inc. (work conducted at the Department of Methodology, London School of Economics and Political Science)
 ---
 
 # Summary
