+++ paper/paper.md
@@ -13,10 +13,10 @@ authors:
 - affiliation: 3
   name: Paul Nulty
   orcid: 0000-0002-7214-4666
-- affiliation: 5
+- affiliation: 4
   name: Adam Obeng
   orcid: 0000-0002-2906-4775
-- affiliation: 4
+- affiliation: 5
   name: Stefan Müller
   orcid: 0000-0002-6315-4125
 - affiliation: 1
@@ -35,9 +35,9 @@ affiliations:
 - index: 3
   name: Centre for Research in Arts, Social Science and Humanities, University of Cambridge
 - index: 4
-  name: Department of Political Science, Trinity College Dublin
-- index: 5
   name: Facebook, Inc. (work conducted at the Department of Methodology, London School of Economics and Political Science)
+- index: 5
+  name: Department of Political Science, Trinity College Dublin
 ---
 
 # Summary
