### 2018-06-17 8d609b7
Tidy up Rd entries
