### 2018-06-23 4fd9534
Add attributes to tokens object after combining - Disallow combining tokens with different concatenator or what values - Join and sort ngrams or skip values
