### 2018-06-23 382a31f
Simplify code
