### 2018-06-23 5629ad8
Follog the best practice in sequence generation
