### 2018-06-25 7ab2b5d
Fix predict.textmodel_wordscores to produce NA for new documents. - Substitute NA with zero before computing document scores - Improve code readability
