### 2018-06-28 be228b1
Set encoding to types when there is no gap or duplication
