### 2018-06-28 0b2a419
Make the type of types explicit
