### 2018-07-02 2dbe5a7
Disallow data.frame without colnames (#1388)
