### 2018-07-03 efcd5ae
Revert "Disallow data.frame without colnames (#1388)"
