### 2018-07-03 d784fd5
Implement new solution to wordscores with NAs
- Omits features from the `wordscores` in a `textmodel_wordscores` class object that were all zero in a training set.
- Implements a new warning when features are found in `newdata` (through `predict.textmodel_wordscores()`) for which no word scores were computed.  These values will not contribute to the computation of the text's predicted document score.
- Keeps the old variable names that @koheiw didn't like because they match the original quantity names from LBG (2003) (for better or worse...)
- Fixes #1380
