### 2018-07-03 ff2185c
Update man pages
