+++ man/head.corpus.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/corpus-methods-base.R
+% Please edit documentation in r/corpus-methods-base.R
 \name{head.corpus}
 \alias{head.corpus}
 \alias{tail.corpus}
