+++ man/types.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/tokens.R
+% Please edit documentation in r/tokens.R
 \name{types}
 \alias{types}
 \title{Get word types from a tokens object}
