+++ man/summary.textmodel_nb.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/textmodel_nb.R
+% Please edit documentation in r/textmodel_nb.R
 \name{summary.textmodel_nb}
 \alias{summary.textmodel_nb}
 \title{summary method for textmodel_nb objects}
