+++ man/replace_dictionary_values.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/dictionaries.R
+% Please edit documentation in r/dictionaries.R
 \name{replace_dictionary_values}
 \alias{replace_dictionary_values}
 \title{Internal function to replace dictionary values}
