+++ man/dfm_tfidf.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/dfm_weight.R
+% Please edit documentation in r/dfm_weight.R
 \name{dfm_tfidf}
 \alias{dfm_tfidf}
 \title{Weight a dfm by \emph{tf-idf}}
