+++ man/tokens_segment.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/tokens_segment.R
+% Please edit documentation in r/tokens_segment.R
 \name{tokens_segment}
 \alias{tokens_segment}
 \title{Segment tokens object by patterns}
