+++ man/reexports.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/spacyr-methods.R, R/stopwords.R, R/utils.R
+% Please edit documentation in r/spacyr-methods.R, r/stopwords.R, r/utils.R
 \docType{import}
 \name{reexports}
 \alias{reexports}
