+++ man/predict.textmodel_wordscores.Rd
@@ -1,12 +1,12 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/textmodel_wordscores.R
+% Please edit documentation in r/textmodel_wordscores.R
 \name{predict.textmodel_wordscores}
 \alias{predict.textmodel_wordscores}
 \title{Predict textmodel_wordscores}
 \usage{
 \method{predict}{textmodel_wordscores}(object, newdata = NULL,
   se.fit = FALSE, interval = c("none", "confidence"), level = 0.95,
-  rescaling = c("none", "lbg", "mv"), force = FALSE, ...)
+  rescaling = c("none", "lbg", "mv"), force = TRUE, ...)
 }
 \arguments{
 \item{object}{a fitted Wordscores textmodel}
@@ -23,7 +23,10 @@
 rescaling; or \code{"mv"} for the rescaling proposed by Martin and Vanberg 
 (2007).  See References.}
 
-\item{force}{make newdata's feature set conformant to the model terms}
+\item{force}{make the feature set of \code{newdata} conform to the model
+terms.  The default of \code{TRUE} means that a fitted model can be applied
+to scale a dfm that doesn't contain a 1:1 match of features in the training
+and prediction data.}
 
 \item{...}{not used}
 }
