+++ man/data_corpus_irishbudget2010.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/data-documentation.R
+% Please edit documentation in r/data-documentation.R
 \docType{data}
 \name{data_corpus_irishbudget2010}
 \alias{data_corpus_irishbudget2010}
