+++ man/as.coefficients_textmodel.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/textmodel-methods.R
+% Please edit documentation in r/textmodel-methods.R
 \name{as.coefficients_textmodel}
 \alias{as.coefficients_textmodel}
 \title{Coerce various objects to coefficients_textmodel
