+++ man/print.statistics_textmodel.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/textmodel-methods.R
+% Please edit documentation in r/textmodel-methods.R
 \name{print.statistics_textmodel}
 \alias{print.statistics_textmodel}
 \title{Implements print methods for textmodel_statistics}
