### 2018-07-03 quanteda/quanteda#1393 @borishejblum
JOSS review
Below are my comments regarding the review items I could not check off for the JOSS submission (so should be easy to fix):

**General checks**
 - License: missing a plain-text GPL-3 licence file in the GitHub repo (only a html file is present). I suggest adding a LICENSE file at the root of the package and adding it to .Rbuildignore to adhere to CRAN policies
 - Version: is it v1.2 or v1.3 ? I am getting confused

**Documentation**
 - Statement of need: goal is clearly stated, but not the targeted audience

**Software paper**
 - References: missing DOIs for all references
 - Statement of need: goal is clearly stated, but not the targeted audience
