### 2018-07-03 #774 @borishejblum
My comments are here https://github.com/quanteda/quanteda/issues/1393
