### 2018-07-04 69ab937
Simplify the code
