+++ man/print.dfm.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in r/dfm-print.R, r/fcm.R
+% Please edit documentation in R/dfm-print.R, R/fcm.R
 \docType{methods}
 \name{print.dfm}
 \alias{print.dfm}
