+++ man/create.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in r/utils.R
+% Please edit documentation in R/utils.R
 \name{create}
 \alias{create}
 \title{Utility function to create a object with new set of attributes}
