+++ man/friendly_class_undefined_message.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in r/utils.R
+% Please edit documentation in R/utils.R
 \name{friendly_class_undefined_message}
 \alias{friendly_class_undefined_message}
 \title{Print friendly object class not defined message}
