+++ man/attributes-set.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in r/utils.R
+% Please edit documentation in R/utils.R
 \name{attributes<-}
 \alias{attributes<-}
 \title{Function extending base::attributes()}
