+++ man/tokens_replace.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in r/tokens_replace.R
+% Please edit documentation in R/tokens_replace.R
 \name{tokens_replace}
 \alias{tokens_replace}
 \title{Replace types in tokens object}
