+++ man/dfm_tolower.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in r/casechange-functions.R
+% Please edit documentation in R/casechange-functions.R
 \name{dfm_tolower}
 \alias{dfm_tolower}
 \alias{dfm_toupper}
