+++ man/tokens_lookup.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in r/tokens_lookup.R
+% Please edit documentation in R/tokens_lookup.R
 \name{tokens_lookup}
 \alias{tokens_lookup}
 \title{Apply a dictionary to a tokens object}
