+++ man/textplot_keyness.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in r/textplot_keyness.R
+% Please edit documentation in R/textplot_keyness.R
 \name{textplot_keyness}
 \alias{textplot_keyness}
 \title{Plot word keyness}
