+++ man/docvars.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in r/docvars.R
+% Please edit documentation in R/docvars.R
 \name{docvars}
 \alias{docvars}
 \alias{docvars<-}
