### 2018-07-04 e697ac1
Revert changes
