+++ R/corpus.R
@@ -159,11 +159,11 @@ corpus.character <- function(x, docnames = NULL,
     
     # replace all hyphens with simple hyphen
     x <- stri_replace_all_regex(x, "\\p{Pd}", "-")
-
+    
     # normalize EOL
     x <- stri_replace_all_fixed(x, "\r\n", "\n") # Windows
     x <- stri_replace_all_fixed(x, "\r", "\n") # Old Macintosh
-
+    
     # name the texts vector
     if (!is.null(docnames)) {
         stopifnot(length(docnames) == length(x))
@@ -174,11 +174,11 @@ corpus.character <- function(x, docnames = NULL,
         # if they previously existed, but got obliterated by a stringi function
         names(x) <- name
     }
-
+    
     # ensure that docnames are unique
     if (any(duplicated(names(x))))
         names(x) <- make.unique(names(x))
-
+    
     # create document-meta-data
     if (is.null(metacorpus$source)) {
         metacorpus$source <- paste(getwd(), "/* ", "on ",  
@@ -188,13 +188,13 @@ corpus.character <- function(x, docnames = NULL,
     if (is.null(metacorpus$created)) {
         metacorpus$created <- date()
     }
-
+    
     # create the documents dframe starting with the texts, using an empty field
     # this saves space if it needs to be separated later
     documents <- data.frame(texts = rep(NA, length(x)),
                             row.names = names(x),
                             check.rows = TRUE, stringsAsFactors = FALSE)
-
+    
     # user-supplied document-level variables (one kind of meta-data)
     if (!is.null(docvars)) {
         if (nrow(docvars) > 0) {
@@ -202,10 +202,10 @@ corpus.character <- function(x, docnames = NULL,
             documents <- cbind(documents, docvars)
         }
     }
-
+    
     # initialize results corpus
     temp <- list()
-
+    
     ## compress and separate texts if compress == TRUE
     # paste delimiters into object to be compressed
     if (compress) {
@@ -218,13 +218,13 @@ corpus.character <- function(x, docnames = NULL,
         # otherwise replace NA placeholder with the actual text
         documents$texts <- x
     }
-
+    
     # build and return the corpus object
     temp <- c(temp, list(documents = documents,
                          metadata = metacorpus,
                          settings = settings(),
                          tokens = NULL))
-
+    
     ## add some elements if compress
     if (compress) {
         temp$docnames <- names(x)
@@ -232,7 +232,7 @@ corpus.character <- function(x, docnames = NULL,
         temp$compression_rate <- utils::object.size(temp$texts) / 
             utils::object.size(unname(x)) * 100
     }
-
+    
     class(temp) <- c("corpus", class(temp))
     if (compress) {
         class(temp) <- c("corpuszip", class(temp))
@@ -250,9 +250,9 @@ corpus.character <- function(x, docnames = NULL,
 #' @export
 corpus.data.frame <- function(x, docid_field = "doc_id", text_field = "text",
                               metacorpus = NULL, compress = FALSE, ...) {
-
+    
     unused_dots(...)
-
+    
     # text_field handling 
     if (length(text_field) != 1)
         stop("text_field must refer to a single column")
@@ -263,12 +263,12 @@ corpus.data.frame <- function(x, docid_field = "doc_id", text_field = "text",
             stop("text_field index refers to an invalid column")
         }
     }
-
+    
     if (!text_field %in% names(x))
         stop("column name ", text_field, " not found")
     if (!is.character(x[[text_field]]))
         stop("text_field must refer to a character mode column")
-
+    
[…]
