### 2018-07-04 15fe8ae
Increment version minor number
