+++ paper/paper.bib
@@ -19,6 +19,7 @@ archivePrefix = "arXiv",
 	Journal = {American Political Science Review},
 	Number = {2},
 	Pages = {311--331},
+	doi = {10.1017/S0003055403000698},
 	Title = {Estimating the policy positions of political actors using words as data},
 	Volume = {97},
 	Year = {2003}}
@@ -30,6 +31,7 @@ archivePrefix = "arXiv",
 	Journal = {American Journal of Political Science},
 	Number = {3},
 	Pages = {705-722},
+	doi = {10.1111/j.1540-5907.2008.00338.x},
 	Title = {A Scaling Model for Estimating Time-Series Party Positions from Texts},
 	Volume = {52},
 	Year = {2008}}
@@ -48,6 +50,7 @@ archivePrefix = "arXiv",
 	Date-Modified = {2014-06-03 09:57:23 +0000},
 	Publisher = {Cambridge University Press},
 	Address = {Cambridge},
+	doi = {10.1017/CBO9780511809071},
 	Title = {Introduction to Information Retrieval},
 	Year = {2008}}
 
@@ -55,6 +58,7 @@ archivePrefix = "arXiv",
   title={Indexing by latent semantic analysis},
   author={Deerwester, Scott and Dumais, Susan T and Furnas, George W and Landauer, Thomas K and Harshman, Richard},
   journal={Journal of the American society for information science},
+  doi = {10.1002/(SICI)1097-4571(199009)41:6%3C391::AID-ASI1%3E3.0.CO;2-9},
   volume={41},
   number={6},
   pages={391},
@@ -64,6 +68,7 @@ archivePrefix = "arXiv",
 @article{youngsoroka2012,
   title={Affective news: The automated coding of sentiment in political texts},
   author={Young, Lori and Soroka, Stuart},
+  doi={10.1080/10584609.2012.671234},
   journal={Political Communication},
   volume={29},
   number={2},
@@ -82,6 +87,7 @@ archivePrefix = "arXiv",
   title={tidytext: Text mining and analysis using tidy data principles in r},
   author={Silge, Julia and Robinson, David},
   journal={The Journal of Open Source Software},
+  doi={10.21105/joss.00037},
   volume={1},
   number={3},
   year={2016}
@@ -107,6 +113,7 @@ archivePrefix = "arXiv",
   title={Text mining infrastructure in R},
   author={Meyer, David and Hornik, Kurt and Feinerer, Ingo},
   journal={Journal of statistical software},
+  doi = {10.18637/jss.v025.i05},
   volume={25},
   number={5},
   pages={1--54},
