### 2018-07-09 bdeed86
add zenodo DOIs
