+++ paper/paper.bib
@@ -80,6 +80,7 @@ archivePrefix = "arXiv",
    AUTHOR  = {Honnibal, Matthew and Montani, Ines},
    TITLE   = {spaCy 2: Natural language understanding with Bloom embeddings, convolutional neural networks and incremental parsing},
    YEAR    = {2017},
+   doi = {10.5281/zenodo.1212304},
    note = {Explosion.ai manuscript}
 }
 
@@ -132,6 +133,7 @@ archivePrefix = "arXiv",
     title = {R package stringi: Character string processing facilities},
     author = {Marek Gagolewski},
     year = {2018},
+    doi = {10.5281/zenodo.1292492},
     url = {http://www.gagolewski.com/software/stringi/},
 }
  
