### 2018-07-09 8514757
Add first names of authors; capitalise titles and journal names
