+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @ARTICLE{PerryBenoit2017,
-   author = {{Perry}, P.~O. and {Benoit}, K.},
+   author = {Perry, Patrick O. and Benoit, Kennet},
     title = "{Scaling Text with the Class Affinity Model}",
   journal = {ArXiv e-prints},
 archivePrefix = "arXiv",
@@ -20,7 +20,7 @@ archivePrefix = "arXiv",
 	Number = {2},
 	Pages = {311--331},
 	doi = {10.1017/S0003055403000698},
-	Title = {Estimating the policy positions of political actors using words as data},
+	Title = {Extracting Policy Positions from Political Texts Using Words as Data},
 	Volume = {97},
 	Year = {2003}}
 	
@@ -30,14 +30,14 @@ archivePrefix = "arXiv",
 	Date-Modified = {2014-06-03 09:57:23 +0000},
 	Journal = {American Journal of Political Science},
 	Number = {3},
-	Pages = {705-722},
+	Pages = {705--722},
 	doi = {10.1111/j.1540-5907.2008.00338.x},
 	Title = {A Scaling Model for Estimating Time-Series Party Positions from Texts},
 	Volume = {52},
 	Year = {2008}}
 
 @book{greenacre1984,
-  title={Theory and applications of correspondence analysis},
+  title={Theory and Applications of Correspondence Analysis},
   author={Greenacre, Michael J},
   year={1984},
   address = {London},
@@ -45,7 +45,7 @@ archivePrefix = "arXiv",
 }
 
 @book{Manningetal2008,
-	Author = {C. D. Manning and P. Raghavan and H. Sch\"{u}tze},
+	Author = {Manning, Christopher D and Raghavan, Prabhakar and Sch\"{u}tze, Hinrich},
 	Date-Added = {2014-06-03 09:57:23 +0000},
 	Date-Modified = {2014-06-03 09:57:23 +0000},
 	Publisher = {Cambridge University Press},
@@ -55,9 +55,9 @@ archivePrefix = "arXiv",
 	Year = {2008}}
 
 @article{deerwester1990,
-  title={Indexing by latent semantic analysis},
+  title={Indexing by Latent Semantic Analysis},
   author={Deerwester, Scott and Dumais, Susan T and Furnas, George W and Landauer, Thomas K and Harshman, Richard},
-  journal={Journal of the American society for information science},
+  journal={Journal of the American Society for Information Science},
   doi = {10.1002/(SICI)1097-4571(199009)41:6%3C391::AID-ASI1%3E3.0.CO;2-9},
   volume={41},
   number={6},
@@ -66,7 +66,7 @@ archivePrefix = "arXiv",
 }
 
 @article{youngsoroka2012,
-  title={Affective news: The automated coding of sentiment in political texts},
+  title={Affective News: The Automated Coding of Sentiment in Political Texts},
   author={Young, Lori and Soroka, Stuart},
   doi={10.1080/10584609.2012.671234},
   journal={Political Communication},
@@ -78,14 +78,14 @@ archivePrefix = "arXiv",
 
 @UNPUBLISHED{spacy2,
    AUTHOR  = {Honnibal, Matthew and Montani, Ines},
-   TITLE   = {spaCy 2: Natural language understanding with Bloom embeddings, convolutional neural networks and incremental parsing},
+   TITLE   = {spaCy 2: Natural Language Understanding with Bloom Embeddings, Convolutional Neural Networks and Incremental Parsing},
    YEAR    = {2017},
    doi = {10.5281/zenodo.1212304},
    note = {Explosion.ai manuscript}
 }
 
 @article{tidytextJOSS,
-  title={tidytext: Text mining and analysis using tidy data principles in r},
[…]
