+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @ARTICLE{PerryBenoit2017,
-   author = {Perry, Patrick O. and Benoit, Kennet},
+   author = {Perry, Patrick O. and Benoit, Kenneth},
     title = "{Scaling Text with the Class Affinity Model}",
   journal = {ArXiv e-prints},
 archivePrefix = "arXiv",
@@ -61,7 +61,7 @@ archivePrefix = "arXiv",
   doi = {10.1002/(SICI)1097-4571(199009)41:6%3C391::AID-ASI1%3E3.0.CO;2-9},
   volume={41},
   number={6},
-  pages={391},
+  pages={391-407},
   year={1990}
 }
 
