### 2018-07-11 2d55bd6
Return original tokens if n=1 and skip=0
