### 2018-07-11 b12439b
Reimplement segment_texts() to support multiple patterns
