+++ R/corpus_segment.R
@@ -127,8 +127,7 @@ corpus_segment.corpus <- function(x, pattern = "##*",
     
     temp <- segment_texts(texts(x), pattern = pattern, valuetype = valuetype, 
                           extract_pattern = extract_pattern, 
-                          pattern_position = pattern_position,
-                          omit_empty = !extract_pattern)
+                          pattern_position = pattern_position)
 
     # create the new corpus
     result <- corpus(temp$texts, docnames = rownames(temp),
@@ -178,9 +177,9 @@ char_segment <- function(x, pattern = "##*",
 
 #' @export
 char_segment.default <- function(x, pattern = "##*",
-                                   valuetype = c("glob", "regex", "fixed"),
-                                   remove_pattern = TRUE,
-                                   pattern_position = c("before", "after")) {
+                                 valuetype = c("glob", "regex", "fixed"),
+                                 remove_pattern = TRUE,
+                                 pattern_position = c("before", "after")) {
     stop(friendly_class_undefined_message(class(x), "char_segment"))
 }
         
@@ -221,14 +220,8 @@ segment_texts <- function(x, pattern = NULL, valuetype = "regex",
     }
     
     if (is.null(pattern)) {
-        
-        # if (what == "tokens") {
-        #     temp <- as.list(tokens(x, ...))
-        # } else 
-        if (what == "sentences") {
-            temp <- as.list(tokens(x, what = "sentence", ...))
-        }
-        
+        if (what == "sentences")
+            x <- as.list(tokens(x, what = "sentence", ...))
     } else {
         
         if (valuetype == "glob") {
@@ -236,61 +229,82 @@ segment_texts <- function(x, pattern = NULL, valuetype = "regex",
             if (!any(stri_detect_charclass(pattern, c("[*?]")))) {
                 valuetype <- "fixed"
             } else {
-                regex <- escape_regex(pattern)
-                regex <- stri_replace_all_fixed(regex, '*', '(\\S*)')
-                regex <- stri_replace_all_fixed(regex, '?', '(\\S)')
-                pattern <- stri_c(regex, collapse = '|')
+                pattern <- escape_regex(pattern)
+                pattern <- stri_replace_all_fixed(pattern, '*', '(\\S*)')
+                pattern <- stri_replace_all_fixed(pattern, '?', '(\\S)')
                 valuetype <- "regex"
             }
         }
         
-        temp <- stri_trim_both(x)
+        x <- stri_trim_both(x)
         if (valuetype == "fixed") {
             if (pattern_position == "after") {
-                temp <- stri_replace_all_fixed(temp, pattern, 
-                                               stri_c(pattern, "\uE000"))
+                x <- stri_replace_all_fixed(x, pattern, 
+                                            stri_c(pattern, "\uE000"),
+                                            vectorize_all = FALSE)
             } else {
-                temp <- stri_replace_all_fixed(temp, pattern, 
-                                               stri_c("\uE000", pattern))
+                x <- stri_replace_all_fixed(x, pattern, 
+                                            stri_c("\uE000", pattern),
+                                            vectorize_all = FALSE)
             }
         } else {
             if (pattern_position == "after") {
-                temp <- stri_replace_all_regex(temp, pattern, "$0\uE000")
+                x <- stri_replace_all_regex(x, pattern, "$0\uE000",
+                                               vectorize_all = FALSE)
             } else {
-                temp <- stri_replace_all_regex(temp, pattern, "\uE000$0")
+                x <- stri_replace_all_regex(x, pattern, "\uE000$0",
+                                               vectorize_all = FALSE)
             }
         }
[…]
