### 2018-07-11 6a89879
Add case_insensitive
