+++ R/corpus_segment.R
@@ -14,6 +14,7 @@
 #' @param x character or \link{corpus} object whose texts will be segmented
 #' @inheritParams pattern
 #' @inheritParams valuetype
+#' @param case_insensitive ignore case when matching, if \code{TRUE}
 #' @param extract_pattern extracts matched patterns from the texts and save in docvars if
 #'   \code{TRUE}
 #' @param pattern_position either \code{"before"} or \code{"after"}, depending
@@ -96,6 +97,7 @@
 #' @export
 corpus_segment <- function(x, pattern = "##*",
                            valuetype = c("glob", "regex", "fixed"),
+                           case_insensitive = TRUE,
                            extract_pattern = TRUE,
                            pattern_position = c("before", "after"),
                            use_docvars = TRUE) {
@@ -105,6 +107,7 @@ corpus_segment <- function(x, pattern = "##*",
 #' @export    
 corpus_segment.default <- function(x, pattern = "##*",
                                   valuetype = c("glob", "regex", "fixed"),
+                                  case_insensitive = TRUE,
                                   extract_pattern = TRUE,
                                   pattern_position = c("before", "after"),
                                   use_docvars = TRUE) {
@@ -114,6 +117,7 @@ corpus_segment.default <- function(x, pattern = "##*",
 #' @export    
 corpus_segment.corpus <- function(x, pattern = "##*",
                                   valuetype = c("glob", "regex", "fixed"),
+                                  case_insensitive = TRUE,
                                   extract_pattern = TRUE,
                                   pattern_position = c("before", "after"),
                                   use_docvars = TRUE) {
@@ -125,9 +129,8 @@ corpus_segment.corpus <- function(x, pattern = "##*",
     commands <- as.character(sys.calls())
     commands <- commands[stri_detect_regex(commands, "segment\\.corpus")]
     
-    temp <- segment_texts(texts(x), pattern = pattern, valuetype = valuetype, 
-                          extract_pattern = extract_pattern, 
-                          pattern_position = pattern_position)
+    temp <- segment_texts(texts(x), pattern, valuetype, case_insensitive,
+                          extract_pattern, pattern_position)
 
     # create the new corpus
     result <- corpus(temp$texts, docnames = rownames(temp),
@@ -170,6 +173,7 @@ corpus_segment.corpus <- function(x, pattern = "##*",
 #' @return \code{char_segment} returns a character vector of segmented texts
 char_segment <- function(x, pattern = "##*",
                          valuetype = c("glob", "regex", "fixed"),
+                         case_insensitive = TRUE,
                          remove_pattern = TRUE,
                          pattern_position = c("before", "after")) {
     UseMethod("char_segment")
@@ -178,6 +182,7 @@ char_segment <- function(x, pattern = "##*",
 #' @export
 char_segment.default <- function(x, pattern = "##*",
                                  valuetype = c("glob", "regex", "fixed"),
+                                 case_insensitive = TRUE,
                                  remove_pattern = TRUE,
                                  pattern_position = c("before", "after")) {
     stop(friendly_class_undefined_message(class(x), "char_segment"))
@@ -186,13 +191,15 @@ char_segment.default <- function(x, pattern = "##*",
 #' @export
 char_segment.character <- function(x, pattern = "##*",
                                    valuetype = c("glob", "regex", "fixed"),
+                                   case_insensitive = TRUE,
                                    remove_pattern = TRUE,
                                    pattern_position = c("before", "after")) {
     
     valuetype <- match.arg(valuetype)
     pattern_position <- match.arg(pattern_position)
     
-    temp <- segment_texts(x, pattern, valuetype, remove_pattern, 
+    temp <- segment_texts(x, pattern, valuetype, 
+                          case_insensitive, remove_pattern, 
                           pattern_position)
[…]
