### 2018-07-11 d6cecd3
Add and restore important tests
