### 2018-07-11 d9a739c
Skip the thread competition test on CI
