### 2018-07-11 377d92a
Update competition test
