### 2018-07-15 d74415d
Update NEWS and DESCRIPTION
