+++ DESCRIPTION
@@ -1,5 +1,5 @@
 Package: quanteda
-Version: 1.3.3
+Version: 1.3.4
 Title: Quantitative Analysis of Textual Data
 Description: A fast, flexible, and comprehensive framework for 
     quantitative text analysis in R.  Provides functionality for corpus management,
