### 2018-07-15 2473d2d
Update CRAN comments
