+++ index.md
@@ -1,13 +1,11 @@
 
-# quanteda: Quantitative Analysis of Textual Data
+quanteda: Quantitative Analysis of Textual Data
+===============================================
 
-## About
+**quanteda** is an R package for managing and analyzing textual data developed by [Kenneth Benoit](http://kenbenoit.net) and other contributors. Initial development was supported by the European Research Council grant ERC-2011-StG 283794-QUANTESS.
 
-An R package for managing and analyzing text, created by [Kenneth
-Benoit](http://kenbenoit.net). Supported by the European Research
-Council grant ERC-2011-StG 283794-QUANTESS.
-
-## How to Install
+How to Install
+--------------
 
 The normal way from CRAN, using your R GUI or
 
@@ -22,44 +20,40 @@ Or for the latest development version:
 devtools::install_github("quanteda/quanteda") 
 ```
 
-Because this compiles some C++ and Fortran source code, you will need to
-have installed the appropriate compilers.
+Because this compiles some C++ and Fortran source code, you will need to have installed the appropriate compilers.
 
-**If you are using a Windows platform**, this means you will need also
-to install the [Rtools](https://CRAN.R-project.org/bin/windows/Rtools/)
-software available from CRAN.
+**If you are using a Windows platform**, this means you will need also to install the [Rtools](https://CRAN.R-project.org/bin/windows/Rtools/) software available from CRAN.
 
-**If you are using macOS**, you should install the [macOS
-tools](https://cran.r-project.org/bin/macosx/tools/), namely the Clang
-6.x compiler and the GNU Fortran compiler (as **quanteda** requires
-gfortran to build).
+**If you are using macOS**, you should install the [macOS tools](https://cran.r-project.org/bin/macosx/tools/), namely the Clang 6.x compiler and the GNU Fortran compiler (as **quanteda** requires gfortran to build).
 
-## How to Use
+System Requirements
+-------------------
 
-See the [quick start
-guide](http://docs.quanteda.io/articles/quickstart.html) to learn how to
-use **quanteda**.
+**quanteda** is cross-platform but we recommend MacOS or Linux as an operating system for their better handling of Unicode. RAM depends on the size and the structure of the textual data to analyze. Usually, a text file of 100MB on disk takes 500MB to 1GB on memory as a tokens object (short texts require more memory than long texts when the total numbers of words are the same).
 
-## Leaving Feedback
+|     |       Minimum       |   Recommended   |
+|-----|:-------------------:|:---------------:|
+| OS  | Windows/MacOS/Linux |   MacOS/Linux   |
+| CPU |        1 core       | 4 cores or more |
+| RAM |         2GB         |  8GB more more  |
+| IDE |                     |     R Studio    |
 
-If you like **quanteda**, please consider leaving [feedback or a
-testimonial here](https://github.com/quanteda/quanteda/issues/461).
+How to Use
+----------
 
-## Contributing
+See the [quick start guide](http://docs.quanteda.io/articles/quickstart.html) to learn how to use **quanteda**.
 
-Contributions in the form of feedback, comments, code, and bug reports
-are most welcome. How to contribute:
+Leaving Feedback
+----------------
 
-  - Fork the source code, modify, and issue a [pull
-    request](https://help.github.com/articles/creating-a-pull-request-from-a-fork/)
-    through the [project GitHub
-    page](https://github.com/quanteda/quanteda). See our [Contributor
-    Code of
-    Conduct](https://github.com/quanteda/quanteda/blob/master/CONDUCT.md)
-    and the all-important **quanteda** [Style
-    Guide](https://github.com/quanteda/quanteda/wiki/Style-guide).
[…]
