+++ index.md
@@ -1,59 +1,106 @@
-
 quanteda: Quantitative Analysis of Textual Data
 ===============================================
 
-**quanteda** is an R package for managing and analyzing textual data developed by [Kenneth Benoit](http://kenbenoit.net) and other contributors. Initial development was supported by the European Research Council grant ERC-2011-StG 283794-QUANTESS.
+**quanteda** is an R package for managing and analyzing textual data
+developed by [Kenneth Benoit](http://kenbenoit.net) and other
+contributors. Its initial development was supported by the European
+Research Council grant ERC-2011-StG 283794-QUANTESS.
 
 How to Install
 --------------
 
 The normal way from CRAN, using your R GUI or
 
-``` r
-install.packages("quanteda") 
-```
+    install.packages("quanteda") 
 
 Or for the latest development version:
 
-``` r
-# devtools package required to install quanteda from Github 
-devtools::install_github("quanteda/quanteda") 
-```
+    # devtools package required to install quanteda from Github 
+    devtools::install_github("quanteda/quanteda") 
 
-Because this compiles some C++ and Fortran source code, you will need to have installed the appropriate compilers.
+Because this compiles some C++ and Fortran source code, you will need to
+have installed the appropriate compilers.
 
-**If you are using a Windows platform**, this means you will need also to install the [Rtools](https://CRAN.R-project.org/bin/windows/Rtools/) software available from CRAN.
+**If you are using a Windows platform**, this means you will need also
+to install the [Rtools](https://CRAN.R-project.org/bin/windows/Rtools/)
+software available from CRAN.
 
-**If you are using macOS**, you should install the [macOS tools](https://cran.r-project.org/bin/macosx/tools/), namely the Clang 6.x compiler and the GNU Fortran compiler (as **quanteda** requires gfortran to build).
+**If you are using macOS**, you should install the [macOS
+tools](https://cran.r-project.org/bin/macosx/tools/), namely the Clang
+6.x compiler and the GNU Fortran compiler (as **quanteda** requires
+gfortran to build).
 
 System Requirements
 -------------------
 
-**quanteda** is cross-platform but we recommend MacOS or Linux as an operating system for their better handling of Unicode. RAM depends on the size and the structure of the textual data to analyze. Usually, a text file of 100MB on disk takes 500MB to 1GB on memory as a tokens object (short texts require more memory than long texts when the total numbers of words are the same).
+**quanteda** is cross-platform but we recommend MacOS or Linux as an
+operating system for their better handling of Unicode. RAM depends on
+the size and the structure of the textual data to analyze. Usually, a
+text file of 100MB on disk takes 500MB to 1GB on memory as a tokens
+object (short texts require more memory than long texts when the total
+numbers of words are the same).
 
-|     |       Minimum       |   Recommended   |
-|-----|:-------------------:|:---------------:|
-| OS  | Windows/MacOS/Linux |   MacOS/Linux   |
-| CPU |        1 core       | 4 cores or more |
-| RAM |         2GB         |  8GB more more  |
-| IDE |                     |     R Studio    |
+<table>
+<thead>
+<tr class="header">
+<th></th>
+<th align="center">Minimum</th>
+<th align="center">Recommended</th>
+</tr>
+</thead>
+<tbody>
+<tr class="odd">
+<td>OS</td>
+<td align="center">Windows/MacOS/Linux</td>
+<td align="center">MacOS/Linux</td>
+</tr>
+<tr class="even">
+<td>CPU</td>
+<td align="center">1 core</td>
+<td align="center">4 cores or more</td>
+</tr>
[…]
