### 2018-07-17 #774 @kbenoit
Just to flag this for the reviewers, in case it makes a difference: Since submission, we have updated the CRAN version to 1.3.4. Looking forward to your reviews.
