### 2018-07-18 6e9b446
Improve normialization of weights
Devision by max value sometimes make other edges very thin. Devision by a 90% percentile value seems to be much more robust.
