### 2018-07-19 07de71f
Change percentile value and document about it
