### 2018-07-20 ffa1874
Repair alternative code suggestion in deprecation warning
