### 2018-08-14 #774 @arfon
> @arfon I am working on this review. Technically I have a conflict of interest according the the JOSS policies, since Ken and I published a JOSS paper together earlier this year. I had assumed that JOSS already knew about this, but I guess not. I do not think this COI should preclude me from submitting my review, but that's your editorial call. What do you think?

Thanks for disclosing this @lmullen. As we have multiple reviewers here I'm happy for you to proceed.
