### 2018-08-15 d53c4e0
Fix counting when ordered = TRUE to solve #1413
