+++ man/dfm_sort.Rd
@@ -4,7 +4,8 @@
 \alias{dfm_sort}
 \title{Sort a dfm by frequency of one or more margins}
 \usage{
-dfm_sort(x, decreasing = TRUE, margin = c("features", "documents", "both"))
+dfm_sort(x, decreasing = TRUE, margin = c("features", "documents",
+  "both"))
 }
 \arguments{
 \item{x}{Document-feature matrix created by \code{\link{dfm}}}
