### 2018-08-18 88ec3f2
Rebuild Rd files - not sure why these were rebuilt in the first place for this PR
