+++ man/textmodel_nb.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/textmodel_nb.R
+% Please edit documentation in r/textmodel_nb.R
 \name{textmodel_nb}
 \alias{textmodel_nb}
 \title{Naive Bayes classifier for texts}
