+++ man/tokens_wordstem.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/wordstem.R
+% Please edit documentation in r/wordstem.R
 \name{tokens_wordstem}
 \alias{tokens_wordstem}
 \alias{char_wordstem}
