+++ man/wordcloud.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/textplot_wordcloud.R
+% Please edit documentation in r/textplot_wordcloud.R
 \name{wordcloud}
 \alias{wordcloud}
 \title{Internal function for textplot_wordcloud}
