+++ man/as.matrix.dist_selection.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/textstat_dist.R
+% Please edit documentation in r/textstat_dist.R
 \name{as.matrix.dist_selection}
 \alias{as.matrix.dist_selection}
 \title{Coerce a dist_selection object to a matrix}
