+++ man/head.dfm.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/dfm-print.R
+% Please edit documentation in r/dfm-print.R
 \name{head.dfm}
 \alias{head.dfm}
 \alias{tail.dfm}
