+++ man/pattern2id.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/pattern2fixed.R
+% Please edit documentation in r/pattern2fixed.R
 \name{pattern2id}
 \alias{pattern2id}
 \alias{pattern2fixed}
