+++ man/set_dfm_slots.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/dfm-methods.R
+% Please edit documentation in r/dfm-methods.R
 \name{set_dfm_slots}
 \alias{set_dfm_slots}
 \title{Set values to a dfm's S4 slots}
