+++ man/convert.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/convert.R
+% Please edit documentation in r/convert.R
 \name{convert}
 \alias{convert}
 \title{Convert a dfm to a non-quanteda format}
