+++ man/as.network.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/textplot_network.R
+% Please edit documentation in r/textplot_network.R
 \name{as.network}
 \alias{as.network}
 \title{redefinition of network::as.network()}
