+++ man/textmodel_affinity.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/textmodel_affinity.R
+% Please edit documentation in r/textmodel_affinity.R
 \name{textmodel_affinity}
 \alias{textmodel_affinity}
 \title{Class affinity maximum likelihood text scaling model}
