+++ man/spacyr-methods.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/spacyr-methods.R
+% Please edit documentation in r/spacyr-methods.R
 \name{spacyr-methods}
 \alias{spacyr-methods}
 \alias{spacy_parse.corpus}
