### 2018-08-18 2fc27f6
Update NEWS and version for #1413 fix
