### 2018-08-27 464bb76
Fix stri_length warnings
