+++ NAMESPACE
@@ -584,6 +584,7 @@ importFrom(stats,resid)
 importFrom(stats,residuals)
 importFrom(stats,rstandard)
 importFrom(stats,sd)
+importFrom(stringi,stri_length)
 importFrom(stringi,stri_trans_tolower)
 importFrom(stringi,stri_trans_toupper)
 importFrom(utils,combn)
