+++ R/corpus_trim.R
@@ -1,0 +1,172 @@
+#' Remove sentences based on their token lengths or a pattern match
+#' 
+#' Removes sentences from a corpus or a character vector shorter than a 
+#' specified length.
+#' @param x \link{corpus} or character object whose sentences will be selected.
+#' @param what units of trimming, \code{"sentences"} or \code{"paragraphs"}, or
+#'   \code{"documents"}
+#' @param min_ntoken,max_ntoken minimum and maximum lengths in word tokens 
+#'   (excluding punctuation)
+#' @param exclude_pattern a \pkg{stringi} regular expression whose match (at the
+#'   sentence level) will be used to exclude sentences
+#' @return a \link{corpus} or character vector equal in length to the input.  If
+#'   the input was a corpus, then the all docvars and metadata are preserved. 
+#'   For documents whose sentences have been removed entirely, a null string
+#'   (\code{""}) will be returned.
+#' @export
+#' @keywords corpus internal
+#' @examples
+#' txt <- c("PAGE 1. This is a single sentence.  Short sentence. Three word sentence.",
+#'          "PAGE 2. Very short! Shorter.",
+#'          "Very long sentence, with multiple parts, separated by commas.  PAGE 3.")
+#' mycorp <- corpus(txt, docvars = data.frame(serial = 1:3))
+#' texts(mycorp)
+#' 
+#' # exclude sentences shorter than 3 tokens
+#' texts(corpus_trim(mycorp, min_ntoken = 3))
+#' # exclude sentences that start with "PAGE <digit(s)>"
+#' texts(corpus_trim(mycorp, exclude_pattern = "^PAGE \\d+"))
+#' 
+corpus_trim <- function(x, what = c("sentences", "paragraphs", "documents"),
+                        min_ntoken = 1, max_ntoken = NULL, 
+                        exclude_pattern = NULL) {
+    UseMethod("corpus_trim")
+}
+
+#' @noRd
+#' @export
+corpus_trim.corpus <- function(x, what = c("sentences", "paragraphs", "documents"),
+                               min_ntoken = 1, max_ntoken = NULL, 
+                               exclude_pattern = NULL) {
+    
+    what <- match.arg(what)
+    if (is.null(max_ntoken)) max_ntoken <- 1e10 
+    
+    # segment corpus
+    if (what != "documents") {
+        temp <- corpus_reshape(x, to = what)
+    } else {
+        temp <- x
+    }
+    
+    # exclude based on lengths
+    length <- ntoken(temp, remove_punct = TRUE)
+    temp <- corpus_subset(temp, length >= min_ntoken & length <= max_ntoken)
+    
+    # exclude based on regular expression match
+    if (!is.null(exclude_pattern)) {
+        temp <- corpus_subset(temp, !stri_detect_regex(texts(temp), 
+                                                       exclude_pattern))
+    }
+    
+    if (what != "documents") {
+        result <- corpus_reshape(temp, to = "documents")
+    } else {
+        result <- temp
+    }
+    
+    return(result)
+}
+
+
+#' @rdname corpus_trim
+#' @keywords character
+#' @export
+#' @examples
+#' # trimming character objects
+#' char_trim(txt, "sentences", min_ntoken = 3)
+#' char_trim(txt, "sentences", exclude_pattern = "sentence\\.")
+char_trim <- function(x, what = c("sentences", "paragraphs", "documents"), 
+                      min_ntoken = 1, max_ntoken = NULL, exclude_pattern = NULL) {
+    UseMethod("char_trim")
+}
+
+#' @noRd
+#' @export
+char_trim.character <- function(x, what = c("sentences", "paragraphs", "documents"), 
+                                min_ntoken = 1, max_ntoken = NULL, exclude_pattern = NULL) {
+    what <- match.arg(what)
[…]

[…]
