### 2018-08-27 e387866
Remove rownames from docvars for tokens and dfm
