+++ tests/testthat/test-docvars.R
@@ -273,6 +273,7 @@ test_that("object always have docvars in the same rows as documents", {
 
 test_that("can assign docvars when value is a dfm (#1417)", {
     mycorp <- corpus(data_char_ukimmig2010)
+
     thedfm <- dfm(mycorp)[, "the"]
     docvars(mycorp) <- thedfm
     expect_identical(
@@ -284,13 +285,13 @@ test_that("can assign docvars when value is a dfm (#1417)", {
     docvars(anddfm) <- anddfm
     expect_identical(
         docvars(anddfm),
-        data.frame(and = as.vector(anddfm), row.names = docnames(mycorp))
+        data.frame(and = as.vector(anddfm), row.names = 1:ndoc(mycorp)) # docnames(mycorp))
     )
 
     toks <- tokens(mycorp)
     docvars(toks) <- anddfm
     expect_identical(
         docvars(toks),
-        data.frame(and = as.vector(anddfm), row.names = docnames(mycorp))
+        data.frame(and = as.vector(anddfm), row.names = 1:ndoc(mycorp)) # docnames(mycorp))
     )
 })
