### 2018-08-27 7b5909b
Increment version, add news for #1417
