+++ NEWS.md
@@ -3,6 +3,7 @@
 ### Bug fixes and stability enhancements
 
 * Fixed a bug causing incorrect counting in `fcm(x, ordered = TRUE)`. (#1413)  Also set the condition that `window` can be of size 1 (formerly the limit was 2 or greater).
+* Fixed deprecation warnings from adding a dfm as docvars, and this now inmports the feature names as docvar names automatically. (related to #1417)
 
 
 # quanteda v1.3.4
