### 2018-08-30 6edc4c3
Call pad_dfm() only when features are not identical
