### 2018-08-30 ddffbda
- Change handing of NULL assignment to preserve data.frame - Add check for the number of rows when x is a data.frame
