### 2018-08-30 868b0fb
Make sure that DFMs from cbind() or rbind() always have docvars in which ndoc() == nrow(), but without rownames
