### 2018-08-30 682f19f
Add check for empty dfm to textmodel_*
