### 2018-08-30 98ef6e6
Automatically create docvars without rownames
