### 2018-08-30 bdd054d
Update Dale-Chall w/extended wordlist
Adds the variants of the D-C words that will produce a match to the variations using fixed matching.
