+++ tests/data_creation/create_wordlists.R
@@ -1,11 +1,16 @@
 makeWordList <- function(filename) {
-    wordList <- readtext(filename, cache = FALSE)@texts
-    wordList <- stringi::stri_replace_all_regex(wordList, "-", "_")
-    wordList <- tokens(wordList, simplify = TRUE)
-    wordList <- stringi::stri_replace_all_regex(wordList, "_", "-")
+    wordList <- readtext::readtext(filename)[["text"]]
+    wordList <- tokens(wordList, remove_hyphens = FALSE, what = "fasterword") %>%
+        as.character() %>%
+        stringi::stri_trim_both() %>%
+        sort()
     wordList
 }
-dalechall    <- makeWordList("~/Dropbox/QUANTESS/quanteda_working_files/readability/Dale-Chall.txt")
-spache    <- makeWordList("~/Dropbox/QUANTESS/quanteda_working_files/readability/Spache.txt")
-wordlists <- list(dalechall = dalechall, spache = spache)
+
+dalechall_old    <- makeWordList("~/Dropbox (Personal)/QUANTESS/quanteda_working_files/readability/Dale-Chall.txt")
+dalechall_extended <- makeWordList("~/Dropbox (Personal)/QUANTESS/quanteda_working_files/readability/Dale-Chall-extended.txt")
+
+
+spache    <- makeWordList("~/Dropbox (Personal)/QUANTESS/quanteda_working_files/readability/Spache.txt")
+wordlists <- list(dalechall = dalechall_extended, spache = spache)
 devtools::use_data(wordlists)
