### 2018-08-30 a54a6af
Update wordlist creation code
