+++ R/tokens_setdiff.R
@@ -1,0 +1,44 @@
+#' Return the set difference of tokens
+#' 
+#' Returns the tokens not found in a character vector.  This is used for
+#' instance in counting the "unfamiliar" words not found in a list, for instance
+#' with the Dale-Chall readability index.
+#' 
+#' @inheritParams pattern
+#' @inheritParams valuetype
+#' @param remove_duplicates if \code{TRUE}, do not return duplicates
+#' @return tokens not found in \code{pattern}
+#' @seealso \code{\link{textstat_readability}}, \code{\link{data_char_wordlist}}
+#' @keywords tokens
+#' @export
+#' @examples 
+#' toks <- tokens(c(d1 = "This sentence is extremely brief.", 
+#'                  d2 = "Two heads are superior to one."), remove_punct = TRUE)
+#' tokens_setdiff(tokens_tolower(toks), pattern = data_char_wordlists$dalechall)
+#' 
+#' txt <- c("a a b c d", "x y c c")
+#' tokens_setdiff(tokens(txt), c("c", "d"))
+#' tokens_setdiff(tokens(txt), c("c", "d"), unique = TRUE)
+tokens_setdiff <- function(x, pattern, valuetype = c("fixed", "glob", "regex"),
+                           unique = FALSE) {
+    UseMethod("tokens_setdiff")
+}
+
+#' @export
+tokens_setdiff.default <- function(x, pattern, valuetype = c("fixed", "glob", "regex"),
+                           unique = FALSE) {
+    stop(friendly_class_undefined_message(class(x), "tokens_setdiff"))
+}
+    
+#' @export
+tokens_setdiff.tokens <- function(x, pattern, valuetype = c("fixed", "glob", "regex"),
+                           unique = FALSE) {
+    valuetype <- match.arg(valuetype)
+    
+    if (valuetype == "fixed") 
+        ret <- lapply(tokens_tolower(x), function(y) y[!(y %in% pattern)])
+    else
+        stop("glob and regex patterns are not yet implemented for this function")
+    
+    if (unique) lapply(ret, unique) else ret
+}
