+++ man/tokens_setdiff.Rd
@@ -1,0 +1,40 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/tokens_setdiff.R
+\name{tokens_setdiff}
+\alias{tokens_setdiff}
+\title{Return the set difference of tokens}
+\usage{
+tokens_setdiff(x, pattern, valuetype = c("fixed", "glob", "regex"),
+  unique = FALSE)
+}
+\arguments{
+\item{pattern}{a character vector, list of character vectors, \link{dictionary},
+\link{collocations}, or \link{dfm}. See \link{pattern} for details.}
+
+\item{valuetype}{the type of pattern matching: \code{"glob"} for 
+"glob"-style wildcard expressions; \code{"regex"} for regular expressions;
+or \code{"fixed"} for exact matching. See \link{valuetype} for details.}
+
+\item{remove_duplicates}{if \code{TRUE}, do not return duplicates}
+}
+\value{
+tokens not found in \code{pattern}
+}
+\description{
+Returns the tokens not found in a character vector.  This is used for
+instance in counting the "unfamiliar" words not found in a list, for instance
+with the Dale-Chall readability index.
+}
+\examples{
+toks <- tokens(c(d1 = "This sentence is extremely brief.", 
+                 d2 = "Two heads are superior to one."), remove_punct = TRUE)
+tokens_setdiff(tokens_tolower(toks), pattern = data_char_wordlists$dalechall)
+
+txt <- c("a a b c d", "x y c c")
+tokens_setdiff(tokens(txt), c("c", "d"))
+tokens_setdiff(tokens(txt), c("c", "d"), unique = TRUE)
+}
+\seealso{
+\code{\link{textstat_readability}}, \code{\link{data_char_wordlist}}
+}
+\keyword{tokens}
