### 2018-08-30 eed85f0
Add checks for empty dfm or fcm
