+++ R/utils.R
@@ -229,7 +229,8 @@ unused_dots <- function(...) {
 #' @param key type of erro message
 #' @keywords internal
 message_error <- function(key = NULL) {
-    msg <- c("dfm_empty" = "dfm must have at least one non-zero feature",
+    msg <- c("dfm_empty" = "dfm must have at least one non-zero value",
+             "fcm_empty" = "fcm must have at least one non-zero value",
              "docvar_mismatch" = "data.frame must have the same number of rows as documents")
     if (is.null(key) || !key %in% names(msg)) {
         return("")
