+++ tests/testthat/test-textstat_lexdiv.R
@@ -79,3 +79,12 @@ test_that("textstat_lexdiv works with a single document dfm (#706)", {
         tolerance = 0.01
     )
 })
+
+test_that("raises error when dfm is empty (#1419)", {
+    
+    mx <- dfm_trim(data_dfm_lbgexample, 1000)
+    expect_error(textstat_lexdiv(mx, c("TTR", "C")),
+                 quanteda:::message_error("dfm_empty"))
+    
+})
+
