### 2018-08-31 386831d
Remove non-tests
