### 2018-08-31 5900838
Fix tokens_setdiff() calls
