### 2018-09-01 4cde243
Update wordlist, add cites
