### 2018-09-01 530b755
Improve wordlist lookup
