+++ R/tokens_setdiff.R
@@ -6,6 +6,7 @@
 #' 
 #' @inheritParams pattern
 #' @inheritParams valuetype
+#' @param case_insensitive ignore case when matching, if \code{TRUE}
 #' @param remove_duplicates if \code{TRUE}, do not return duplicates
 #' @return tokens not found in \code{pattern}
 #' @seealso \code{\link{textstat_readability}}, \code{\link{data_char_wordlist}}
@@ -20,25 +21,40 @@
 #' tokens_setdiff(tokens(txt), c("c", "d"))
 #' tokens_setdiff(tokens(txt), c("c", "d"), unique = TRUE)
 tokens_setdiff <- function(x, pattern, valuetype = c("fixed", "glob", "regex"),
+                           case_insensitive = FALSE, remove_possessives = FALSE,
                            unique = FALSE) {
     UseMethod("tokens_setdiff")
 }
 
 #' @export
 tokens_setdiff.default <- function(x, pattern, valuetype = c("fixed", "glob", "regex"),
-                           unique = FALSE) {
+                                   case_insensitive = FALSE, remove_possessives = FALSE,
+                                   unique = FALSE) {
     stop(friendly_class_undefined_message(class(x), "tokens_setdiff"))
 }
     
 #' @export
 tokens_setdiff.tokens <- function(x, pattern, valuetype = c("fixed", "glob", "regex"),
-                           unique = FALSE) {
-    valuetype <- match.arg(valuetype)
+                                  case_insensitive = FALSE, remove_possessives = FALSE,
+                                  unique = FALSE) {
     
-    if (valuetype == "fixed") 
+    valuetype <- match.arg(valuetype)
+    pattern <- as.character(pattern)
+    
+    if (valuetype == "fixed") {
+        if (case_insensitive) {
+            x <- tokens_tolower(x)
+            pattern <- char_tolower(pattern)
+        }
+        if (remove_possessives) {
+            x <- stringi::stri_replace_last_charclass(x, "['’]", "")  
+            pattern <- stringi::stri_replace_last_charclass(pattern, "['’]", "")
+        }
         ret <- lapply(tokens_tolower(x), function(y) y[!(y %in% pattern)])
-    else
+        
+    } else {
         stop("glob and regex patterns are not yet implemented for this function")
+    }
     
     if (unique) lapply(ret, unique) else ret
 }
