### 2018-09-01 ad359e6
Fix #1420
- tokens(x, what = "fasterword", remove_separators = TRUE)` so that it correctly splits words separated by `\n` and `\t` characters.
- However: the handling of tokens using the `remove_*` options is still inconsistent depending on the `what` value.
