### 2018-09-03 1dc7ed5
Add tests from original Dale and Chall (1948)
