+++ tests/testthat/test-textstat_readability.R
@@ -50,54 +50,74 @@ test_that("readability works as koRpus", {
 })
 
 test_that("Test Dale-Chall readability", {
-    # from Jeanne Chall and Edgar Dale’s Readability Revisited: 
-    # The New Dale-Chall Readability Formula featured the following text samples, 
-    # with the difficult words not found on their new word list underlined (pp. 135-140).
-    # see http://www.impact-information.com/scales.pdf
-    
-    dc1 <- 'One morning Toad sat in bed.
-    “I have many things to do,” he said.
-    “I will write them all down on a list so that I can remember them.”
-    Toad wrote on a piece of paper: A list of things to do today.
-    Then he wrote: Wake up.
-    “I have done that,” said Toad, and he crossed it out.'
-    expect_identical(
-        lengths(tokens_setdiff(tokens(dc1, remove_punct = TRUE), char_tolower(data_char_wordlists$dalechall))),
-        c(text1 = 0L)
-    )
-    toks <- tokens(dc1, remove_punct = TRUE)
-    textstat_readability(dc1, "Dale.Chall.old")
-        
-    dc2 <- 'You said you didn’t want it,” said Thelma. “And anyhow, I don’t want to sell it now.” “Why not?” said Frances.
-    “Well,” said Thelma, “it is a very good tea set. It is plastic that does not break.
-    It has pretty red flowers on it. It has all the cups and saucers.
-    It has the sugar bowl and the cream pitcher and the teapot.
-    It is almost new, and I think it cost a lot of money.”
-    “I have two dollars and seventeen cents,” said Frances.
-    “That’s a lot of money.”
-    “I don’t know,” said Thelma. “If I sell you..."'
-    toks <- tokens(dc2, remove_punct = TRUE)
-    textstat_readability(dc2, "Dale.Chall.old")
-    
+    # from Dale, Edgar, and Jeanne S Chall. 1948. “A Formula for Predicting
+    # Readability: Instructions.” Educational Research Bulletin 27(2): 37–54.
 
-    # Readability Data
-    # Number of Words in Sample ........................100 
-    # Number of Whole Sentences...........................8 
-    # Number of Unfamiliar Words..........................3 
-    # Cloze Score .......................................53 
-    # Reading Level ..................................... 3
-    dc3 <- "Once upon a time a very small witch was walking in the woods. The cold wind was blowing the dry leaves all around her. The little witch was frantically searching for a house for the winter. She could not find one. Suddenly a piece of orange paper, blown by the wind, landed at her feet. She picked it up. The little witch looked closely at the paper and then she said, “I shall make myself a little house from this piece of orange paper.”
-    She folded the paper in half. then she took her scissors (she always carried a pair..."
-    expect_equal(textstat_readability(dc3, measure = c("Dale.Chall.old"))$Dale.Chall.old, 3)
+    dc1 <- "A happy, useful life - that's what you want for your baby, isn't it? And because
+a healthy mind and body are so necessary to happiness and long life, you must
+do all you can to get your baby off to a good start. There is much you can do
+while he is still a baby to lay the foundation for good health and good health
+habits. Many things affect your baby's health. One was the state of your own
+health during pregnancy, and the special care your doctor gave you before the
+baby was born. Other things important to your child's health are food,
+clothes, baths, sleep, and habit training. A baby needs a clean, happy place
+to live, and he must be kept from having any sickness that can be prevented."
+    unfamiliar_words <- tokens_setdiff(tokens(dc1, remove_punct = TRUE), 
[…]
