### 2018-09-03 8e7043b
Increment version and add NEWS for #1410
