### 2018-09-03 2cbf0f3
Use tokens_setdiff for Spache
