### 2018-09-04 c1cd525
Remove `tokens_setdiff()`
(since `tokens_remove()` does the same thing.)
