+++ tests/testthat/test-textstat_readability.R
@@ -62,12 +62,12 @@ health during pregnancy, and the special care your doctor gave you before the
 baby was born. Other things important to your child's health are food,
 clothes, baths, sleep, and habit training. A baby needs a clean, happy place
 to live, and he must be kept from having any sickness that can be prevented."
-    unfamiliar_words <- tokens_setdiff(tokens(dc1, remove_punct = TRUE), 
+    unfamiliar_words <- tokens_remove(tokens(dc1, remove_punct = TRUE), 
                                        pattern = char_tolower(data_char_wordlists$dalechall),
                                        case_insensitive = TRUE) %>%
         as.character()
     expect_identical(
-        unfamiliar_words,
+        unique(unfamiliar_words),
         c("necessary", "foundation", "affect", "pregnancy", "special", "prevented")
     )
     expect_equivalent(ntoken(dc1, remove_punct = TRUE), 132)
@@ -83,12 +83,12 @@ Your doctor will tell you that your baby should have this protection before
 his first birthday. Six months after the last injection of toxoid, the
 physician may test your baby to see if another dose of toxoid is necessary.
 Before the child enters school an extra shot of toxoid is often given."
-    unfamiliar_words <- tokens_setdiff(tokens(dc2, remove_punct = TRUE), 
+    unfamiliar_words <- tokens_remove(tokens(dc2, remove_punct = TRUE), 
                                        pattern = char_tolower(data_char_wordlists$dalechall),
                                        case_insensitive = TRUE) %>%
         as.character()
     expect_identical(
-        unique(unfamiliar_words),
+        unique(char_tolower(unfamiliar_words)),
         c("diphtheria", "diseases", "treatment", "prevention", "disease", 
           "immunization", "physicians", "usually", "injections", "doses", 
           "toxoid", "protection", "injection", "physician", "dose", "necessary")
@@ -106,12 +106,12 @@ dishes, his toys. The baby's hands may carry germs from soiled objects to his
 mouth. Kissing is one way of spreading TB as well as other germs. Tuberculosis
 of the bones or joints or of certain organs of the body besides the lungs can
 come to the bottle-fed baby in milk which has not been pasteurized or boiled."
-    unfamiliar_words <- tokens_setdiff(tokens(dc3, remove_punct = TRUE), 
+    unfamiliar_words <- tokens_remove(tokens(dc3, remove_punct = TRUE), 
                                        pattern = char_tolower(data_char_wordlists$dalechall),
                                        case_insensitive = TRUE) %>%
         as.character()
     expect_identical(
-        unique(unfamiliar_words),
+        unique(char_tolower(unfamiliar_words)),
         c("germs", "tuberculosis", "spray", "moisture", "active", "germ-filled",
           "objects", "tb", "joints", "lungs", "bottle-fed", "pasteurized")
     )
