+++ NAMESPACE
@@ -356,8 +356,6 @@ S3method(tokens_replace,tokens)
 S3method(tokens_segment,tokens)
 S3method(tokens_select,default)
 S3method(tokens_select,tokens)
-S3method(tokens_setdiff,default)
-S3method(tokens_setdiff,tokens)
 S3method(tokens_skipgrams,default)
 S3method(tokens_skipgrams,tokens)
 S3method(tokens_subset,default)
@@ -509,7 +507,6 @@ export(tokens_remove)
 export(tokens_replace)
 export(tokens_segment)
 export(tokens_select)
-export(tokens_setdiff)
 export(tokens_skipgrams)
 export(tokens_subset)
 export(tokens_tolower)
