+++ DESCRIPTION
@@ -166,7 +166,6 @@ Collate:
     'tokens_replace.R'
     'tokens_segment.R'
     'tokens_select.R'
-    'tokens_setdiff.R'
     'tokens_subset.R'
     'utils.R'
     'wordstem.R'

[…]
