### 2018-09-06 9b0f98e
Add wordlist data to core repo
