### 2018-09-06 d9db335
Add intermediate argument to textstat_readability
Outputs intermediate quantities used in computing readability statistics. For use with the [sophistication package](https://github.com/kbenoit/sophistication/).
