+++ tests/data_creation/create_wordlists.R
@@ -1,3 +1,5 @@
+library("quanteda")
+
 makewordlist <- function(filename) {
     wordlist <- readtext::readtext(filename)[["text"]] %>%
         stringi::stri_split_regex(pattern = "\\p{WHITE_SPACE}+", simplify = TRUE) %>%
