### 2018-09-06 6d77a94
Update NEWS for `textstat_readability(..., intermediate)`
