+++ NEWS.md
@@ -8,6 +8,9 @@
 * Add error checking for functions taking dfm inputs in case a dfm has empty features (#1419).
 * For `textstat_readability()`, fixed a bug in Dale-Chall-based measures and in the Spache word list measure.  These were caused by an incorrect lookup mechanism but also by limited implementation of the wordlists.  The new wordlists include all of the variations called for in the original measures, but using fast fixed matching. (#1410)
 
+### New Features
+
+* Added new argument `intermediate` to `textstat_readability(x, measure, intermediate = FALSE)`, which if `TRUE` returns intermediate quantities used in the computation of readability statistics.  Useful for verification or direct use of the intermediate quantities.
 
 # quanteda v1.3.4
 
