### 2018-09-10 #774 @lmullen
By the end of the day Friday.

On Mon, Sep 10, 2018 at 7:15 AM Arfon Smith <notifications@github.com>
wrote:

> 👋 @alexgarciac <https://github.com/alexgarciac> @lmullen
> <https://github.com/lmullen> - when do you think you might be able to
> complete your review by?
>
