### 2018-09-14 #774 @lmullen
quanteda is an impressive package, well thought out and well implemented. The package is well documented and the test suite is quite extensive.

On the test suite:

- The test suite emits numerous warnings. Some of these can obviously be ignored by comparing what is being tested to the warning message. Others, it is not clear why a warning message is being given.
- I found at least three test failures, all relating to `textplot_wordcloud`, with all suggested packages installed.
- I installed all the suggested packages listed in the DESCRIPTION. Should the `austin` package also be listed as a suggested package? Should the `wordcloud` package?

JOSS specific comments:

- The repository does not have "a plain-text LICENSE file" so I have not checked that box. However, following R package convention the license is indicated in the DESCRIPTION, and I don't think there is a need for a separate file.
- The version in the repository (1.3.7) is several minor releases ahead of the published CRAN version (1.3.4).
- The JOSS paper (at least, based on the PDF proof and on the BibTeX file) does not have DOIs for a number of the articles, though a spot-check confirms that those article have been assigned DOI.

Comments on the text of the paper:

- "on entirely sparse operations" should be "entirely on sparse operations"
- "quanteda serializes tokens into integers." Is serialization what is happening, or would "hashes" be a better verb?
- "make use of the sparseness document-feature matrices": word missing?
- Describing families of functions as "textplot" (in quotation marks) seems odd to me. It might be clearer that those are prefixes for the functions by writing something like `textplot_*`. That format is used later on in the design section.
- "These materials help beginner users to understand how to use these functions for basic operations, and expert users to combine the functions for advanced text processing and analysis" should be "These materials help beginner users understand how to use these functions for basic operations and expert users combine the functions for advanced text processing and analysis"
- "process large textual data that are difficult for other R packages (such as computing distances or scoring collocations)": This is a misplaced modifier. The "such as" describes analytical computations, not "large textual data." I'd recommend a revision like "perform analyses on large textual data ... such as"
- The readtext package should have a citation like the others, even though it is by the same authors.
- There are a number of small typographical errors in the PDF proof that should be checked.
