### 2018-09-17 #774 @kbenoit
Thanks @arfon, that's easily done and we are happy to work on the other comments and the statement of need. I can also eliminate the local CHECK failures that are caused by tests that try to open graphics devices. Should I consider the reviews complete now, and proceed with the revision?
