### 2018-09-17 #774 @arfon
> Should I consider the reviews complete now, and proceed with the revision?

Yes, please go ahead.
