### 2018-09-17 d3d9732
First working sparse similarity function
