+++ man/textstat_simil2.Rd
@@ -1,0 +1,76 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/textstat_simil2.R
+\name{textstat_simil2}
+\alias{textstat_simil2}
+\title{Similarity and distance computation between documents or features}
+\usage{
+textstat_simil2(x, selection = NULL, margin = c("documents",
+  "features"), method = c("cosine", "correlation"), upper = FALSE,
+  diag = FALSE, min_simil = 0)
+}
+\arguments{
+\item{x}{a \link{dfm} object}
+
+\item{selection}{a valid index for document or feature names from \code{x},
+to be selected for comparison}
+
+\item{margin}{identifies the margin of the dfm on which similarity or
+difference will be computed:  \code{"documents"} for documents or 
+\code{"features"} for word/term features}
+
+\item{method}{method the similarity or distance measure to be used; see
+Details}
+
+\item{upper}{whether the upper triangle of the symmetric \eqn{V \times V} 
+matrix is recorded}
+
+\item{diag}{whether the diagonal of the distance matrix should be recorded}
+}
+\value{
+\code{textstat_simil2} and \code{textstat_dist} return
+  \code{\link{dist}} class objects if selection is \code{NULL}, otherwise, a
+  matrix is returned matching distances to the documents or features
+  identified in the selection.
+}
+\description{
+These functions compute matrixes of distances and similarities between 
+documents or features from a \code{\link{dfm}} and return a 
+\code{\link[stats]{dist}} object (or a matrix if specific targets are
+selected).  They are fast and robust because they operate directly on the sparse
+\link{dfm} objects.
+}
+\details{
+\code{textstat_simil2} options are: \code{"correlation"} (default), 
+  \code{"cosine"}, \code{"jaccard"}, \code{"ejaccard"}, \code{"dice"},
+  \code{"edice"}, \code{"simple matching"}, \code{"hamann"}, and 
+  \code{"faith"}.
+}
+\note{
+If you want to compute similarity on a "normalized" dfm object 
+  (controlling for variable document lengths, for methods such as correlation
+  for which different document lengths matter), then wrap the input dfm in 
+  \code{\link{dfm_weight}(x, "prop")}.
+}
+\examples{
+# similarities for documents
+pres_dfm <- dfm(data_corpus_inaugural, remove_punct = TRUE, remove = stopwords("english"))
+(s1 <- textstat_simil2(pres_dfm, method = "cosine", margin = "documents"))
+as.matrix(s1)
+as.list(s1)
+
+# similarities for for specific documents
+textstat_simil2(pres_dfm, "2017-Trump", margin = "documents")
+textstat_simil2(pres_dfm, "2017-Trump", method = "cosine", margin = "documents")
+textstat_simil2(pres_dfm, c("2009-Obama" , "2013-Obama"), margin = "documents")
+
+# compute some term similarities
+s2 <- textstat_simil2(pres_dfm, c("fair", "health", "terror"), method = "cosine", 
+                      margin = "features")
+head(as.matrix(s2), 10)
+as.list(s2, n = 8)
+
+}
+\seealso{
+\code{\link{textstat_dist}}, \code{\link{as.list.dist}},
+  \code{\link{dist}}
+}

[…]
