### 2018-09-17 91ddb0a
- Transpose dfm in the R side - Add correlation measure - Return Sparase Symmetric Matrix
