+++ src/simil_mt.cpp
@@ -1,66 +1,95 @@
 #include "armadillo.h"
 #include "quanteda.h"
+#include "dev.h"
 using namespace quanteda;
 using std::pow;
 using std::exp;
 using std::sqrt;
 using std::log;
 
-
-struct cosine_similarity : public Worker {
+double simil_cosine(arma::colvec& col_i, 
+                     arma::colvec& col_j) {
     
-    const arma::sp_mat& amat; // input
+    double simil = dot(col_i, col_j) / (sqrt(accu(pow(col_i, 2))) * sqrt(accu(pow(col_j, 2))));
+    return(simil);
+}
+
+double simil_correlation(arma::colvec& col_i, 
+                          arma::colvec& col_j) {
+    
+    double simil = as_scalar(cor(col_i, col_j));
+    return(simil);
+}
+
+struct similarity : public Worker {
+    
+    const arma::sp_mat& mat; // input
     Triplets& simil_tri; // output
-    const std::vector<unsigned int>& selection;
-    double limit;
+    const int method;
+    const std::vector<unsigned int>& target;
+    const double limit;
     arma::uword nrow, ncol;
     
-    // initialize from Rcpp input and output matrixes 
-    cosine_similarity(const arma::sp_mat& amat_, Triplets& simil_tri_, 
-                      std::vector<unsigned int>& selection_,
-                      double limit_, arma::uword nrow_, arma::uword ncol_) :
-                      amat(amat_), simil_tri(simil_tri_), selection(selection_),
-                      limit(limit_), nrow(nrow_), ncol(ncol_) {}
+    similarity(const arma::sp_mat& mat_, Triplets& simil_tri_,
+               int method_, std::vector<unsigned int>& target_,
+               double limit_, arma::uword nrow_, arma::uword ncol_) :
+               method(method_), mat(mat_), simil_tri(simil_tri_), target(target_),
+               limit(limit_), nrow(nrow_), ncol(ncol_) {}
     
-    // function call operator that work for the specified range (begin/end)
     void operator()(std::size_t begin, std::size_t end) {
-        bool upper = selection.size() < nrow;
+        double simil = 0;
+        bool symm = target.size() == ncol;
+        arma::colvec col_zero = arma::zeros<arma::colvec>(nrow);
         for (std::size_t i = begin; i < end; i++) {
-            arma::colvec col_i = amat.col(i) + arma::zeros<arma::colvec>(ncol);
-            for (auto s : selection) {
-                std::size_t j = s - 1;
-                if (!upper && j >= i) continue;
-                arma::colvec col_j = amat.col(j) + arma::zeros<arma::colvec>(ncol);
-                double simil = dot(col_i, col_j) / (sqrt(accu(pow(col_i, 2))) * sqrt(accu(pow(col_j, 2))));
-                //Rcout << "i=" << i << " j=" << j << " simil=" << simil << "\n";
-                if (s >= limit) {
-                    Triplet mat_triplet = std::make_tuple(i, j, simil);
-                    simil_tri.push_back(mat_triplet);
+            arma::colvec col_i = mat.col(i) + col_zero;
+            std::size_t j;
+            for (auto s : target) {
+                j = s - 1;
+                //Rcout << "i=" << i << " j=" << j << "\n";
+                if (symm && j > i) continue;
+                arma::colvec col_j = mat.col(j) + col_zero;
+                switch (method){
+                    case 1:
+                        simil = simil_cosine(col_i, col_j);
+                        break;
+                    case 2:
[…]
