+++ R/RcppExports.R
@@ -49,7 +49,7 @@ qatd_cpp_kwic <- function(texts_, types_, words_, window) {
     .Call(`_quanteda_qatd_cpp_kwic`, texts_, types_, words_, window)
 }
 
-qatd_cpp_similarity <- function(mat, method, target_, limit = 0) {
+qatd_cpp_similarity <- function(mat, method, target_, limit = -1.0) {
     .Call(`_quanteda_qatd_cpp_similarity`, mat, method, target_, limit)
 }
 
