### 2018-09-18 e15f257
Improve performance using grow_by()
