+++ tests/benchmarks/benchmark_textstat_smil.R
@@ -8,6 +8,7 @@ mt <- dfm(data_corpus_inaugural)
 microbenchmark::microbenchmark(
     textstat_simil2(mt, margin = "features"),
     textstat_simil2(mt, margin = "features", min_simil = 0.5),
+    textstat_simil2(mt, margin = "features", min_simil = 0.8),
     textstat_simil(mt, margin = "features", method = "cosine"),
     times = 10
 )
@@ -31,6 +32,7 @@ testthat::expect_equal(
 microbenchmark::microbenchmark(
     textstat_simil2(mt, margin = "features", method = "correlation"),
     textstat_simil2(mt, margin = "features", method = "correlation", min_simil = 0.5),
+    textstat_simil2(mt, margin = "features", method = "correlation", min_simil = 0.8),
     textstat_simil(mt, margin = "features", method = "correlation"),
     times = 10
 )
