### 2018-09-18 bf9d016
Improve perforance for #1210 - Change the way to extract coluns - Create output matrix in the C++ code
