### 2018-09-18 5311532
Further performance improvement - Ignore zeros regardless of simil_limit - Compute magniuted beforehand
