### 2018-09-19 1b8f2fa
Add rank criteria
