+++ R/RcppExports.R
@@ -49,8 +49,8 @@ qatd_cpp_kwic <- function(texts_, types_, words_, window) {
     .Call(`_quanteda_qatd_cpp_kwic`, texts_, types_, words_, window)
 }
 
-qatd_cpp_similarity <- function(mt, method, target_, limit = -1.0) {
-    .Call(`_quanteda_qatd_cpp_similarity`, mt, method, target_, limit)
+qatd_cpp_similarity <- function(mt, method, target_, rank, limit = -1.0) {
+    .Call(`_quanteda_qatd_cpp_similarity`, mt, method, target_, rank, limit)
 }
 
 qatd_cpp_tokens_compound <- function(texts_, comps_, types_, delim_, join) {
