### 2018-09-19 828e7ef
Fix symmetric matrix
