### 2018-09-19 quanteda/quanteda#1427 @kbenoit
Fix local test failures from textplot_wordcloud
These come from `test-textplot.R`, and cause annoying local failures (but not on CI or CRAN where they are skipped). Can we avoid this while still running the tests? Perhaps by not printing the output object?

```
* checking tests ...
  Running ‘testthat.R’ ERROR
Last 13 lines of output:

    |
    |                                                                      |   0%
    |
    |======================================================================| 100%
  ══ testthat results  ═══════════════════════════════════════════════════════════════════════════════════════════
  OK: 3346 SKIPPED: 3 FAILED: 4
  1. Failure: summary.character works with data.frames containing character (#1285) (@test-summary.R#70)
  2. Failure: test textplot_wordcloud comparison works (@test-textplot.R#120)
  3. Failure: test textplot_wordcloud comparison works (@test-textplot.R#123)
  4. Failure: test textplot_wordcloud comparison works (@test-textplot.R#126)

  Error: testthat unit tests failed
  Execution halted
* checking for unstated dependencies in vignettes ... [62s/57s]
 OK
* checking package vignettes in ‘inst/doc’ ... OK
* checking re-building of vignette outputs ... OK
* DONE
Status: 1 ERROR
checking tests ... ERROR
  Running ‘testthat.R’ [62s/57s]
Last 13 lines of output:

    |
    |                                                                      |   0%
    |
    |======================================================================| 100%
  ══ testthat results  ═══════════════════════════════════════════════════════════════════════════════════════════
  OK: 3346 SKIPPED: 3 FAILED: 4
  1. Failure: summary.character works with data.frames containing character (#1285) (@test-summary.R#70)
  2. Failure: test textplot_wordcloud comparison works (@test-textplot.R#120)
  3. Failure: test textplot_wordcloud comparison works (@test-textplot.R#123)
  4. Failure: test textplot_wordcloud comparison works (@test-textplot.R#126)

  Error: testthat unit tests failed
  Execution halted

See
  ‘/Users/kbenoit/Dropbox (Personal)/GitHub/quanteda/quanteda.Rcheck/00check.log’
for details.

R CMD check results
1 error  | 0 warnings | 0 notes

R CMD check succeeded
```
