### 2018-09-19 52641e6
Use nth_element() instead of sort()
