### 2018-09-19 quanteda/quanteda#1430 @kbenoit
Eliminate warnings in tests
Most just need to be wrapped in `suppressWarnings()`. This was one of the JOSS reviewer comments.
