### 2018-09-19 1a0342d
Address @lmullen JOSS review comments
From https://github.com/openjournals/joss-reviews/issues/774#issuecomment-421512862

- Adds `LICENSE` to `.Rbuildignore`
- Adds package citations for **readtext** and **spacyr**
- Fixes awkward or misleading sentences
- Updates a few URLs that have been updated since initial submission.
- Fixes small typos found in the pdf.
