+++ paper/paper.bib
@@ -145,4 +145,18 @@ archivePrefix = "arXiv",
     url = {https://CRAN.R-project.org/package=RcppParallel},
   }
 
-
+@Manual{readtext,
+    title = {readtext: Import and handling for plain and formatted text files},
+    author = {Benoit, Kenneth and Adam Obeng},
+    year = {2018},
+    note = {R package version 0.71},
+    url = {https://CRAN.R-project.org/package=readtext},
+}
+  
+@Manual{spacyr,
+    title = {spacyr: Wrapper to the {spaCy} {NLP} Library},
+    author = {Benoit, Kenneth and Akitaka Matsuo},
+    year = {2018},
+    note = {R package version 0.9.91},
+    url = {https://CRAN.R-project.org/package=spacyr},
+}
