+++ paper/paper.md
@@ -23,6 +23,7 @@ authors:
   name: Akitaka Matsuo
   orcid: 0000-0002-3323-6330
 date: "4 May 2018"
+output: pdf_document
 bibliography: paper.bib
 tags:
 - text mining
@@ -33,9 +34,11 @@ affiliations:
 - index: 2
   name: De Beers Inc.
 - index: 3
-  name: Centre for Research in Arts, Social Science and Humanities, University of Cambridge
+  name: Centre for Research in Arts, Social Science and Humanities, University of
+    Cambridge
 - index: 4
-  name: Facebook, Inc. (work conducted at the Department of Methodology, London School of Economics and Political Science)
+  name: Facebook, Inc. (work conducted at the Department of Methodology, London School
+    of Economics and Political Science)
 - index: 5
   name: Department of Political Science, Trinity College Dublin
 ---
@@ -47,7 +50,7 @@ natural language processing tasks such as corpus management, tokenization,
 analysis, and visualization. It has extensive functions for applying dictionary
 analysis, exploring texts using keywords-in-context, computing document and
 feature similarities, and discovering multi-word expressions through collocation
-scoring. Based on entirely sparse operations, it provides highly efficient
+scoring. Based entirely on sparse operations, it provides highly efficient
 methods for compiling document-feature matrices and for manipulating these or
 using them in further quantitative analysis. Using C++ and multi-threading
 extensively, **quanteda** is also considerably faster and more efficient than
@@ -68,7 +71,7 @@ conditions or combinations of document-level variables.
 **quanteda** is principally designed to allow users a fast and convenient method
 to construct a document-feature matrix from a corpus with an ability to perform
 the most common natural language processing tasks such as tokenizing, stemming,
-forming n-grams, selecting and weighting features.  With these functions, users
+forming n-grams, and selecting and weighting features.  With these functions, users
 can easily remove stop words and stem words in numerous languages, select words
 in a dictionary, and convert frequency counts into weights, for instance using 
 _tf-idf_ scoring.
@@ -76,14 +79,14 @@ _tf-idf_ scoring.
 Using the ICU library in the **stringi** package [@stringi] for text processing,
 **quanteda** correctly handles Unicode character sets for regular expression
 matching and detecting word boundaries for tokenization. Once texts are tokenized,
-**quanteda** serializes tokens into integers to increase processing speed while
+**quanteda** maps tokens to a hash table of integers to increase processing speed while
 reducing memory usage. Many of the text processing functions are parallelized
 using the Intel TBB library via the **RcppParallel** package [@RcppParallel].
 
 # Models and textual statistics
 
 **quanteda** is especially suited to research because it was designed from the
-outset for the social scientific analysis of textual data. Its "textmodel"
+outset for the social scientific analysis of textual data. Its `textmodel_*`
 functions provide native, highly efficient implementations of several text
 analytic scaling methods, such as Wordscores [@lbg:2003], Wordfish
 [@SlapinProksch2008], class affinity scaling [@PerryBenoit2017], and
@@ -97,14 +100,14 @@ Sentiment Dictionary [@youngsoroka2012].
 In addition to models, the package provides a variety of text statistics, 
 such as frequency analysis, "keyness", lexical diversity, readability, and
 similarity and distance of documents or features. These make use of the
-sparseness document-feature matrices -- often over 90% sparse -- and parallelism
+sparsity of document-feature matrices -- often over 90% sparse -- and parallelism
 for efficient, fast computation. **quanteda**  also provides methods for statistically
 scoring collocations, useful in identifying multi-word expressions.
 
 # Text visualization
 
 The package provides extensive methods for visualizing textual analyses, via its
-family of "textplot" functions. These are typically designed to take another
[…]
