### 2018-09-19 #774 @kbenoit
Please see our revised submission in https://github.com/quanteda/quanteda/pull/1431 in the branch `joss-review-response`.

@lmullen you are right that the local tests fail (on macOS) anyway for `textplot_wordcloud()` and this is due to the tests trying to open a graphics device. It does not occur on CI or CRAN, and we will fix this slightly annoying result soon. (https://github.com/quanteda/quanteda/issues/1427)

The warnings from the tests are usually tests of our own warnings, but we will try to clean these up too (https://github.com/quanteda/quanteda/issues/1430).

@alexgarciac we are eager to receive your comments as well, if you have any not already covered by the other two reviewers and @arfon.
