### 2018-09-19 7e410bd
Update version and NEWS
