### 2018-09-20 bb95ff7
Eliminate warnings from tests Eliminate local CHECK error messages from textplot_wordcloud()
- Fixes #1427
- Fixes #1430
