+++ tests/testthat/test-textplot.R
@@ -106,9 +106,7 @@ test_that("test textplot_wordcloud works for dfm objects", {
 })
 
 test_that("test textplot_wordcloud comparison works", {
-    skip_on_travis()
-    skip_on_cran()
-    skip_on_os('linux')
+    skip("skipped because in CHECK mode, fails to open graphics devices")
     testcorp <- corpus_reshape(corpus(data_char_sampletext))
     set.seed(1)
     docvars(testcorp, "label") <- sample(c("A", "B"), size = ndoc(testcorp), replace = TRUE)
@@ -134,7 +132,6 @@ test_that("test textplot_wordcloud comparison works", {
 })
 
 test_that("test textplot_wordcloud raise deprecation message", {
-    
     mt <- dfm(data_corpus_inaugural[1:5])
     mt <- dfm_trim(mt, min_termfreq = 10)
     expect_warning(textplot_wordcloud(mt, min.freq = 10), 'min.freq is deprecated')
@@ -161,7 +158,8 @@ test_that("test textplot_scale1d wordfish in the most basic way", {
 test_that("test textplot_scale1d wordscores in the most basic way", {
     mt <- dfm(data_corpus_irishbudget2010)
     ws <- textmodel_wordscores(mt, c(rep(NA, 4), -1, 1, rep(NA, 8)))
-    pr <- predict(ws, mt, force = TRUE)
+    pr <- suppressWarnings(predict(ws, mt, force = TRUE))
+    
     expect_false(identical(textplot_scale1d(pr, sort = TRUE), 
                            textplot_scale1d(pr, sort = FALSE)))
     expect_silent(textplot_scale1d(pr, sort = TRUE, groups = docvars(data_corpus_irishbudget2010, "party")))
@@ -181,7 +179,7 @@ test_that("test textplot_scale1d wordscores in the most basic way", {
         "This margin can only be run on a predicted wordscores object"
     )
     expect_error(
-        textplot_scale1d(predict(ws), margin = "features"),
+        suppressWarnings(textplot_scale1d(predict(ws), margin = "features")),
         "This margin can only be run on a fitted wordscores object"
     )
 })
