### 2018-09-20 quanteda/quanteda#1430 @kbenoit
I got most of these in bb95ff7 but there are some that I'd prefer @koheiw to address. These are:

```
✔ | 34   8   | test pattern2fixed.R
───────────────────────────────────────────────────────────────────────────────────────────────────────
test-pattern2fixed.R:148: warning: regex2fixed and regex2id still work
argument is not an atomic vector; coercing

test-pattern2fixed.R:148: warning: regex2fixed and regex2id still work
argument is not an atomic vector; coercing

test-pattern2fixed.R:150: warning: regex2fixed and regex2id still work
argument is not an atomic vector; coercing

test-pattern2fixed.R:150: warning: regex2fixed and regex2id still work
argument is not an atomic vector; coercing

test-pattern2fixed.R:153: warning: regex2fixed and regex2id still work
argument is not an atomic vector; coercing

test-pattern2fixed.R:153: warning: regex2fixed and regex2id still work
argument is not an atomic vector; coercing

test-pattern2fixed.R:155: warning: regex2fixed and regex2id still work
argument is not an atomic vector; coercing

test-pattern2fixed.R:155: warning: regex2fixed and regex2id still work
argument is not an atomic vector; coercing
```

It appears to be what happens when you send a list, instead of a vector, to `stringi::stri_stri_trans_nfc()`. Probably can solve it through an `unlist()`.
