### 2018-09-20 eede51e
Supprot all other similarity and disntace measures
