### 2018-09-20 e50595c
Expose docs for corpus_trim/char_trim
