### 2018-09-21 85403f4
update test-textplot.R for mac compatibility
