### 2018-09-23 a49571d
Change http://docs.quanteda.io to https://quanteda.io
