### 2018-09-24 dd5a3ce
Rewrite performance para with allegedly mispaced modifer 😉
