### 2018-09-24 1489646
Minor cosmetic edits to bib file
