+++ paper/paper.bib
@@ -85,7 +85,7 @@ archivePrefix = "arXiv",
 }
 
 @article{tidytextJOSS,
-  title={tidytext: Text Mining and Analysis Using tidy Data Principles in R},
+  title={tidytext: Text Mining and Analysis Using Tidy Data Principles in R},
   author={Silge, Julia and Robinson, David},
   journal={The Journal of Open Source Software},
   doi={10.21105/joss.00037},
@@ -146,7 +146,7 @@ archivePrefix = "arXiv",
   }
 
 @Manual{readtext,
-    title = {readtext: Import and handling for plain and formatted text files},
+    title = {readtext: Import and Handling for Plain and Formatted Text Files},
     author = {Benoit, Kenneth and Adam Obeng},
     year = {2018},
     note = {R package version 0.71},
