### 2018-09-25 c584d26
Change "http://docs.quanteda.io" to "https://quanteda.io"
