+++ docs/articles/design.Rmd
@@ -15,7 +15,7 @@ knitr::opts_chunk$set(
 )
 ```
 
-See http://docs.quanteda.io for additional tutorials, examples, and general documentation.
+See https://quanteda.io for additional tutorials, examples, and general documentation.
 
 # Introduction
 
