### 2018-09-25 f35fb8e
Update code for textplot_wordcloud()
