### 2018-09-26 77a429c
Fix accessor to ws predicted objects
