### 2018-09-26 8f9d5db
Update version and NEWS for #1440
