### 2018-09-26 960065e
Fix more https and docs.quanteda.io references
