### 2018-09-27 e19bdab
Move code from dev-simil-test
