+++ src/simil_mt.cpp
@@ -11,6 +11,9 @@ double magni_correlation(arma::mat& col) {
     return(as_scalar(stddev(col, 1)));
 }
 
+bool condition;
+arma::uvec nz2;
+
 struct magnitude : public Worker {
     
     const arma::sp_mat& mt; // input
@@ -18,7 +21,7 @@ struct magnitude : public Worker {
     const int method;
     
     magnitude(const arma::sp_mat& mt_, DoubleParams& magni_, int method_) :
-              mt(mt_), magni(magni_), method(method_) {}
+        mt(mt_), magni(magni_), method(method_) {}
     
     void operator()(std::size_t begin, std::size_t end) {
         
@@ -29,21 +32,36 @@ struct magnitude : public Worker {
             col = arma::mat(mt.col(i));
             //Rcout << "i=" << i << "\n";
             switch (method) {
-                case 1:
-                    magni[i] = magni_cosine(col);
-                    break;
-                case 2:
-                    magni[i] = magni_correlation(col);
-                    break;
-                default:
-                    magni[i] = 0;
+            case 1:
+                magni[i] = magni_cosine(col);
+                break;
+            case 2:
+                magni[i] = magni_correlation(col);
+                break;
+            default:
+                magni[i] = 0;
             }
         }
     }
 };
 
-double simil_cosine(arma::mat& col_i, arma::mat& col_j, double magni_i, double magni_j) {
+double simil_cosine(const arma::colvec& col_i, const arma::colvec& col_j, 
+                    const double& magni_i, const double& magni_j) {
     return dot(col_i, col_j) / (magni_i * magni_j);
+}
+
+double simil_cosine(const arma::sp_mat& col_i, const arma::sp_mat& col_j, 
+                    const double& magni_i, const double& magni_j) {
+    return dot(col_i, col_j) / (magni_i * magni_j);
+}
+
+double simil_cosine(const arma::colvec& col_i, const arma::colvec& col_j, arma::uvec& nz, 
+                    const double& magni_i, const double& magni_j) {
+    double p = 0;
+    for (arma::uvec::iterator it = nz.begin(); it != nz.end(); ++it) {
+        p += col_i[*it] * col_j[*it];
+    }
+    return p / (magni_i * magni_j);
 }
 
 double simil_correlation(arma::mat& col_i, arma::mat& col_j, double magni_i, double magni_j) {
@@ -91,7 +109,21 @@ double dist_canberra(arma::mat& col_i, arma::mat& col_j) {
 }
 
 double dist_minkowski(arma::mat& col_i, arma::mat& col_j, double order = 1) {
-    return(pow(accu(pow(abs(col_i) - abs(col_j), order)), 1 / order));
+    return pow(accu(pow(abs(col_i) - abs(col_j), order)), 1 / order);
+}
+
+arma::uvec find_nonzero(const arma::sp_mat& mt, int col) {
+    std::vector<int> i;
+    //std::vector<int> i(mt.n_rows, 0);
+    i.reserve(mt.n_rows);
+    for (arma::sp_mat::const_col_iterator it = mt.begin_col(col); it != mt.end_col(col); ++it) {
+        i.push_back(it.row());
+        //Rcout << it.row() << "\n";
+        //i[it.col()] = 1;
+    }
+    //return(arma::ind2sub(arma::size(mt.n_rows, 1), arma::conv_to<arma::uvec>::from(i)));
+    return(arma::conv_to<arma::uvec>::from(i));
+    //return arma::conv_to<arma::uvec>::from(i);
 }
 
 struct similarity : public Worker {
@@ -109,8 +141,8 @@ struct similarity : public Worker {
     similarity(const arma::sp_mat& mt_, Triplets& simil_tri_, const DoubleParams& magni_,
[…]
