### 2018-09-27 a28f1f8
Incorporate faster code
