+++ src/simil_mt.cpp
@@ -4,12 +4,15 @@
 using namespace quanteda;
 
 double magni_cosine(arma::mat& col) {
-    return(sqrt(accu(pow(col, 2))));
+    return(sqrt((double)accu(pow(col, 2))));
 }
 
 double magni_correlation(arma::mat& col) {
     return(as_scalar(stddev(col, 1)));
 }
+
+bool condition;
+arma::uvec nz2;
 
 struct magnitude : public Worker {
     
@@ -18,7 +21,7 @@ struct magnitude : public Worker {
     const int method;
     
     magnitude(const arma::sp_mat& mt_, DoubleParams& magni_, int method_) :
-              mt(mt_), magni(magni_), method(method_) {}
+        mt(mt_), magni(magni_), method(method_) {}
     
     void operator()(std::size_t begin, std::size_t end) {
         
@@ -29,21 +32,39 @@ struct magnitude : public Worker {
             col = arma::mat(mt.col(i));
             //Rcout << "i=" << i << "\n";
             switch (method) {
-                case 1:
-                    magni[i] = magni_cosine(col);
-                    break;
-                case 2:
-                    magni[i] = magni_correlation(col);
-                    break;
-                default:
-                    magni[i] = 0;
+            case 1:
+                magni[i] = magni_cosine(col);
+                break;
+            case 2:
+                magni[i] = magni_correlation(col);
+                break;
+            default:
+                magni[i] = 0;
             }
         }
     }
 };
 
-double simil_cosine(arma::mat& col_i, arma::mat& col_j, double magni_i, double magni_j) {
-    return dot(col_i, col_j) / (magni_i * magni_j);
+arma::uvec find_nonzero(const arma::sp_mat& mt, int col) {
+    std::vector<int> nz;
+    nz.reserve(mt.n_rows);
+    for (arma::sp_mat::const_col_iterator it = mt.begin_col(col); it != mt.end_col(col); ++it) {
+        nz.push_back(it.row());
+    }
+    return(arma::conv_to<arma::uvec>::from(nz));
+}
+
+double simil_cosine(const arma::colvec& col_i, const arma::colvec& col_j, arma::uvec& nz, 
+                    double magni_i, double magni_j) {
+    double p = 0;
+    for (arma::uvec::iterator it = nz.begin(); it != nz.end(); ++it) {
+        p += col_i[*it] * col_j[*it];
+    }
+    return p / (magni_i * magni_j);
+}
+
+double simil_cosine(const arma::colvec& col_i, const arma::colvec& col_j, double magni_i, double magni_j) { 
+       return dot(col_i, col_j) / (magni_i * magni_j);
 }
 
 double simil_correlation(arma::mat& col_i, arma::mat& col_j, double magni_i, double magni_j) {
@@ -91,8 +112,9 @@ double dist_canberra(arma::mat& col_i, arma::mat& col_j) {
 }
 
 double dist_minkowski(arma::mat& col_i, arma::mat& col_j, double order = 1) {
-    return(pow(accu(pow(abs(col_i) - abs(col_j), order)), 1 / order));
+    return pow(accu(pow(abs(col_i) - abs(col_j), order)), 1 / order);
 }
+
 
 struct similarity : public Worker {
     
@@ -109,8 +131,8 @@ struct similarity : public Worker {
     similarity(const arma::sp_mat& mt_, Triplets& simil_tri_, const DoubleParams& magni_,
                const int method_, const std::vector<unsigned int>& target_, 
                const unsigned int rank_, const double limit_, const bool symm_, const double weight_) :
-               mt(mt_), simil_tri(simil_tri_), magni(magni_), 
[…]
