### 2018-09-27 e1886f3
Allow numeric and logical vector for selection
