### 2018-09-28 265c065
Swich to sparse magnitude computation
