### 2018-09-28 a9409ee
User linear algebra to for cosine
