+++ R/RcppExports.R
@@ -49,8 +49,12 @@ qatd_cpp_kwic <- function(texts_, types_, words_, window) {
     .Call(`_quanteda_qatd_cpp_kwic`, texts_, types_, words_, window)
 }
 
-qatd_cpp_similarity <- function(mt, method, target_, rank, limit = -1.0, weight = 1.0) {
-    .Call(`_quanteda_qatd_cpp_similarity`, mt, method, target_, rank, limit, weight)
+qatd_cpp_similarity_linear <- function(mt, method, target_, rank, limit = -1.0, weight = 1.0, condition_ = FALSE) {
+    .Call(`_quanteda_qatd_cpp_similarity_linear`, mt, method, target_, rank, limit, weight, condition_)
+}
+
+qatd_cpp_similarity <- function(mt, method, target_, rank, limit = -1.0, weight = 1.0, condition_ = FALSE) {
+    .Call(`_quanteda_qatd_cpp_similarity`, mt, method, target_, rank, limit, weight, condition_)
 }
 
 qatd_cpp_tokens_compound <- function(texts_, comps_, types_, delim_, join) {
