+++ src/RcppExports.cpp
@@ -169,9 +169,9 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
-// qatd_cpp_similarity
-S4 qatd_cpp_similarity(const arma::sp_mat& mt, const int method, const IntegerVector target_, unsigned int rank, double limit, const double weight);
-RcppExport SEXP _quanteda_qatd_cpp_similarity(SEXP mtSEXP, SEXP methodSEXP, SEXP target_SEXP, SEXP rankSEXP, SEXP limitSEXP, SEXP weightSEXP) {
+// qatd_cpp_similarity_linear
+S4 qatd_cpp_similarity_linear(const arma::sp_mat& mt, const int method, const IntegerVector target_, unsigned int rank, double limit, const double weight, bool condition_);
+RcppExport SEXP _quanteda_qatd_cpp_similarity_linear(SEXP mtSEXP, SEXP methodSEXP, SEXP target_SEXP, SEXP rankSEXP, SEXP limitSEXP, SEXP weightSEXP, SEXP condition_SEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
@@ -181,7 +181,25 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< unsigned int >::type rank(rankSEXP);
     Rcpp::traits::input_parameter< double >::type limit(limitSEXP);
     Rcpp::traits::input_parameter< const double >::type weight(weightSEXP);
-    rcpp_result_gen = Rcpp::wrap(qatd_cpp_similarity(mt, method, target_, rank, limit, weight));
+    Rcpp::traits::input_parameter< bool >::type condition_(condition_SEXP);
+    rcpp_result_gen = Rcpp::wrap(qatd_cpp_similarity_linear(mt, method, target_, rank, limit, weight, condition_));
+    return rcpp_result_gen;
+END_RCPP
+}
+// qatd_cpp_similarity
+S4 qatd_cpp_similarity(const arma::sp_mat& mt, const int method, const IntegerVector target_, unsigned int rank, double limit, const double weight, bool condition_);
+RcppExport SEXP _quanteda_qatd_cpp_similarity(SEXP mtSEXP, SEXP methodSEXP, SEXP target_SEXP, SEXP rankSEXP, SEXP limitSEXP, SEXP weightSEXP, SEXP condition_SEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< const arma::sp_mat& >::type mt(mtSEXP);
+    Rcpp::traits::input_parameter< const int >::type method(methodSEXP);
+    Rcpp::traits::input_parameter< const IntegerVector >::type target_(target_SEXP);
+    Rcpp::traits::input_parameter< unsigned int >::type rank(rankSEXP);
+    Rcpp::traits::input_parameter< double >::type limit(limitSEXP);
+    Rcpp::traits::input_parameter< const double >::type weight(weightSEXP);
+    Rcpp::traits::input_parameter< bool >::type condition_(condition_SEXP);
+    rcpp_result_gen = Rcpp::wrap(qatd_cpp_similarity(mt, method, target_, rank, limit, weight, condition_));
     return rcpp_result_gen;
 END_RCPP
 }
[…]
