### 2018-09-29 14bb6f7
Integrate two functions
