+++ src/simil_mt.cpp
@@ -4,230 +4,242 @@
 using namespace quanteda;
 using namespace arma;
 
-double magni_cosine(const sp_mat& mt, int i) {
-    sp_mat::const_col_iterator it = mt.begin_col(i);
-    sp_mat::const_col_iterator it_end = mt.end_col(i);
-    double sq_sum = 0;
-    for(; it != it_end; ++it) {
-        sq_sum += (*it) * (*it);
-    }
-    return std::sqrt(sq_sum);
+std::vector<double> to_vector(const sp_mat& mt) {
+    return conv_to< std::vector<double> >::from(mat(mt));
 }
 
-double magni_correlation(const sp_mat& mt, int i) {
-    sp_mat::const_col_iterator it = mt.begin_col(i);
-    sp_mat::const_col_iterator it_end = mt.end_col(i);
-    double n = mt.n_rows;
-    if(n == 0)
-        return 0.0;
-    double sum = 0;
-    double sq_sum = 0;
-    for(; it != it_end; ++it) {
-        sum += (*it);
-        sq_sum += (*it) * (*it);
-    }
-    double mean = sum / n;
-    double var = sq_sum / n - mean * mean;
-    return std::sqrt(var);
+std::vector<double> to_vector(const rowvec& v) {
+    return conv_to< std::vector<double> >::from(v);
 }
 
-bool condition;
-
-struct magnitude : public Worker {
+double get_limit(std::vector<double> simils, const unsigned int rank, double limit) {
     
-    const sp_mat& mt; // input
-    DoubleParams& magni; // output
+    if (simils.size() > rank) {
+        std::nth_element(simils.begin(), simils.begin() + rank - 1, simils.end(),
+                         std::greater<double>());
+        if (limit < simils[rank - 1])
+            limit = simils[rank - 1];
+    }
+    return limit;
+}
+
+rowvec stddev(const sp_mat& mt, const int norm_type) {
+    rowvec v(mt.n_cols);
+    for (uword i = 0; i < mt.n_cols; i++) {
+        v[i] = stddev(colvec(mt.col(i)), norm_type);
+    }
+    return(v);
+}
+
+rowvec mean(const sp_mat& mt) {
+    rowvec v(mt.n_cols);
+    for (uword i = 0; i < mt.n_cols; i++) {
+        v[i] = mean(colvec(mt.col(i)));
+    }
+    return(v);
+}
+
+struct similarity_linear : public Worker {
+    
+    const arma::sp_mat& mt1; // input
+    const arma::sp_mat& mt2; // input
+    Triplets& simil_tri; // output
+    const rowvec& square;
+    const rowvec& center;
     const int method;
+    const std::vector<unsigned int>& target;
+    const unsigned int rank;
+    double limit;
+    const bool symm;
     
-    magnitude(const sp_mat& mt_, DoubleParams& magni_, int method_) :
-        mt(mt_), magni(magni_), method(method_) {}
+    similarity_linear(const sp_mat& mt1_, const sp_mat& mt2_, Triplets& simil_tri_, 
+                      const rowvec& square_, const rowvec& center_,
+                      const int method_, const std::vector<unsigned int>& target_, 
+                      const unsigned int rank_, double limit_, const bool symm_) :
+        mt1(mt1_), mt2(mt2_), simil_tri(simil_tri_), square(square_), center(center_), 
+        method(method_), target(target_), rank(rank_), limit(limit_), symm(symm_) {}
     
     void operator()(std::size_t begin, std::size_t end) {
         
-        for (std::size_t i = begin; i < end; i++) {
-            //Rcout << "i=" << i << "\n";
-            switch (method) {
-            case 1:
-                magni[i] = magni_cosine(mt, i);
-                break;
[…]
