### 2018-09-29 ef206d0
Add converter to S4 matrix
