### 2018-09-29 d9d6414
Remove reserve as it does not imporove speed at all
