+++ src/RcppExports.cpp
@@ -170,25 +170,24 @@ BEGIN_RCPP
 END_RCPP
 }
 // qatd_cpp_similarity_linear
-S4 qatd_cpp_similarity_linear(const arma::sp_mat& mt, const int method, const IntegerVector target_, unsigned int rank, double limit, const double weight, bool condition_);
-RcppExport SEXP _quanteda_qatd_cpp_similarity_linear(SEXP mtSEXP, SEXP methodSEXP, SEXP target_SEXP, SEXP rankSEXP, SEXP limitSEXP, SEXP weightSEXP, SEXP condition_SEXP) {
+S4 qatd_cpp_similarity_linear(const arma::sp_mat& mt, const int method, const IntegerVector targets_, unsigned int rank, double limit, bool condition);
+RcppExport SEXP _quanteda_qatd_cpp_similarity_linear(SEXP mtSEXP, SEXP methodSEXP, SEXP targets_SEXP, SEXP rankSEXP, SEXP limitSEXP, SEXP conditionSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
     Rcpp::traits::input_parameter< const arma::sp_mat& >::type mt(mtSEXP);
     Rcpp::traits::input_parameter< const int >::type method(methodSEXP);
-    Rcpp::traits::input_parameter< const IntegerVector >::type target_(target_SEXP);
+    Rcpp::traits::input_parameter< const IntegerVector >::type targets_(targets_SEXP);
     Rcpp::traits::input_parameter< unsigned int >::type rank(rankSEXP);
     Rcpp::traits::input_parameter< double >::type limit(limitSEXP);
-    Rcpp::traits::input_parameter< const double >::type weight(weightSEXP);
-    Rcpp::traits::input_parameter< bool >::type condition_(condition_SEXP);
-    rcpp_result_gen = Rcpp::wrap(qatd_cpp_similarity_linear(mt, method, target_, rank, limit, weight, condition_));
+    Rcpp::traits::input_parameter< bool >::type condition(conditionSEXP);
+    rcpp_result_gen = Rcpp::wrap(qatd_cpp_similarity_linear(mt, method, targets_, rank, limit, condition));
     return rcpp_result_gen;
 END_RCPP
 }
 // qatd_cpp_similarity
-S4 qatd_cpp_similarity(const arma::sp_mat& mt, const int method, const IntegerVector target_, unsigned int rank, double limit, const double weight, bool condition_);
-RcppExport SEXP _quanteda_qatd_cpp_similarity(SEXP mtSEXP, SEXP methodSEXP, SEXP target_SEXP, SEXP rankSEXP, SEXP limitSEXP, SEXP weightSEXP, SEXP condition_SEXP) {
+S4 qatd_cpp_similarity(const arma::sp_mat& mt, const int method, const IntegerVector target_, unsigned int rank, double limit, const double weight);
+RcppExport SEXP _quanteda_qatd_cpp_similarity(SEXP mtSEXP, SEXP methodSEXP, SEXP target_SEXP, SEXP rankSEXP, SEXP limitSEXP, SEXP weightSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
@@ -198,8 +197,7 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< unsigned int >::type rank(rankSEXP);
[…]
