### 2018-09-29 8158298
Enable categorical similarity measures
