### 2018-09-29 d31f44f
Intialize with fixed size
