+++ tests/testthat/test-textstat_simil2.R
@@ -1,0 +1,308 @@
+context("test textstat_simil2.R")
+
+test_mt <- dfm(corpus_subset(data_corpus_inaugural, Year > 1980), remove = stopwords("english"),
+               stem = TRUE, verbose = FALSE)
+
+# correlation
+test_that("test textstat_simil2 method = correlation on documents", {
+    skip_if_not_installed("proxy")
+    s1 <- as.matrix(textstat_simil2(test_mt, method = "correlation", margin = "documents"))
+    s2 <- as.matrix(proxy::simil(as.matrix(test_mt), by_rows = TRUE, diag = TRUE))
+    expect_equal(s1, s2, tolerance = 0.001)
+    
+    s3 <- as.matrix(textstat_simil2(test_mt, "1985-Reagan", method = "correlation", margin = "documents"))
+    s4 <- as.matrix(proxy::simil(as.matrix(test_mt), as.matrix(test_mt["1985-Reagan",]), by_rows = TRUE, diag = TRUE))
+    expect_equal(as.numeric(s3), as.numeric(s4), tolerance = 0.001)
+})
+
+test_that("test textstat_simil2 method = correlation on features)", {
+    skip_if_not_installed("proxy")
+    s1 <- as.matrix(textstat_simil2(test_mt[,1:100], method = "correlation", margin = "features"))
+    s2 <- as.matrix(proxy::simil(as.matrix(test_mt[,1:100]), 
+                                 method = "correlation", by_rows = FALSE, diag = TRUE))
+    expect_equal(s1, s2, tolerance = 0.001)
+    
+    s3 <- as.matrix(textstat_simil2(test_mt, "soviet", method = "correlation", margin = "features"))
+    s4 <- as.matrix(proxy::simil(as.matrix(test_mt), as.matrix(test_mt[,"soviet"]), 
+                                 method = "correlation", by_rows = FALSE, diag = TRUE))
+    expect_equal(as.numeric(s3), as.numeric(s4), tolerance = 0.001)
+})
+
+test_that("test textstat_simil2 method = cosine on documents", {
+    skip_if_not_installed("proxy")
+    s1 <- as.matrix(textstat_simil2(test_mt, method = "cosine", margin = "documents"))
+    s2 <- as.matrix(proxy::simil(as.matrix(test_mt), 
+                                 method = "cosine", by_rows = TRUE, diag = TRUE))
+    expect_equal(s1, s2, tolerance = 0.001)
+    
+    s3 <- as.matrix(textstat_simil2(test_mt, "1985-Reagan", method = "cosine", margin = "documents"))
+    s4 <- as.matrix(proxy::simil(as.matrix(test_mt), as.matrix(test_mt["1985-Reagan",]), 
+                                 method = "cosine", by_rows = TRUE, diag = TRUE))
+    expect_equal(as.numeric(s3), as.numeric(s4), tolerance = 0.001)
+})
+
+test_that("test textstat_simil2 method = cosine on features", {
+    skip_if_not_installed("proxy")
+    s1 <- as.matrix(textstat_simil2(test_mt[,1:100], method = "cosine", margin = "features"))
+    s2 <- as.matrix(proxy::simil(as.matrix(test_mt[,1:100]), 
+                                 method = "cosine", by_rows = FALSE, diag = TRUE))
+    expect_equal(s1, s2, tolerance = 0.001)
+    
+    s3 <- as.matrix(textstat_simil2(test_mt, "soviet", method = "cosine", margin = "features"))
+    s4 <- as.matrix(proxy::simil(as.matrix(test_mt), as.matrix(test_mt[,"soviet"]), 
+                                 method = "cosine", by_rows = FALSE, diag = TRUE))
[…]
