### 2018-10-01 a3cf8c0
Correct spelling of hamman
