+++ src/simil_mt.cpp
@@ -133,7 +133,7 @@ double simil_edice(colvec& col_i, colvec& col_j, double weight = 1) {
     return (2 * e) / (accu(pow(col_i, weight)) + accu(pow(col_j, weight)));
 }
 
-double simil_hamann(colvec& col_i, colvec& col_j, double weight = 1) {
+double simil_hamman(colvec& col_i, colvec& col_j, double weight = 1) {
     double e = accu(col_i == col_j);
     double u = col_i.n_rows - e;
     return (e - (u * weight)) / (e + u);
@@ -211,7 +211,7 @@ struct similarity : public Worker {
                     simil = simil_edice(col_i, col_j, weight);
                     break;
                 case 3:
-                    simil = simil_hamann(col_i, col_j, weight);
+                    simil = simil_hamman(col_i, col_j, weight);
                     break;
                 case 4:
                     simil = simil_faith(col_i, col_j);
