+++ R/textstat_simil2.R
@@ -1,28 +1,29 @@
 #' Similarity and distance computation between documents or features
-#' 
-#' These functions compute matrixes of distances and similarities between 
-#' documents or features from a \code{\link{dfm}} and return a 
+#'
+#' These functions compute matrixes of distances and similarities between
+#' documents or features from a \code{\link{dfm}} and return a
 #' \code{\link[stats]{dist}} object (or a matrix if specific targets are
-#' selected).  They are fast and robust because they operate directly on the sparse
-#' \link{dfm} objects.
+#' selected).  They are fast and robust because they operate directly on the
+#' sparse \link{dfm} objects.
 #' @param x a \link{dfm} object
 #' @param selection a valid index for document or feature names from \code{x},
 #'   to be selected for comparison
 #' @param margin identifies the margin of the dfm on which similarity or
-#'   difference will be computed:  \code{"documents"} for documents or 
+#'   difference will be computed:  \code{"documents"} for documents or
 #'   \code{"features"} for word/term features
 #' @param method method the similarity or distance measure to be used; see
 #'   Details
-#' @param upper  whether the upper triangle of the symmetric \eqn{V \times V} 
-#'   matrix is recorded
-#' @param diag whether the diagonal of the distance matrix should be recorded
-#' @details \code{textstat_simil2} options are: \code{"correlation"} (default), 
+#' @param rank an integer value specifying top-n most similar documents or features
+#'   to be recorded.
+#' @param min_simil minimum similarity value to be recoded.
+#' @param condition boolean flag used only in developemnt
+#' @details \code{textstat_simil2} options are: \code{"correlation"} (default),
 #'   \code{"cosine"}, \code{"jaccard"}, \code{"ejaccard"}, \code{"dice"},
-#'   \code{"edice"}, \code{"simple matching"}, \code{"hamman"}, and 
+#'   \code{"edice"}, \code{"simple matching"}, \code{"hamman"}, and
 #'   \code{"faith"}.
-#' @note If you want to compute similarity on a "normalized" dfm object 
+#' @note If you want to compute similarity on a "normalized" dfm object
 #'   (controlling for variable document lengths, for methods such as correlation
-#'   for which different document lengths matter), then wrap the input dfm in 
+#'   for which different document lengths matter), then wrap the input dfm in
 #'   \code{\link{dfm_weight}(x, "prop")}.
 #' @return \code{textstat_simil2} and \code{textstat_dist} return
 #'   \code{\link{dist}} class objects if selection is \code{NULL}, otherwise, a
@@ -37,14 +38,14 @@
 #' (s1 <- textstat_simil2(pres_dfm, method = "cosine", margin = "documents"))
 #' as.matrix(s1)
 #' as.list(s1)
-#' 
+#'
 #' # similarities for for specific documents
 #' textstat_simil2(pres_dfm, "2017-Trump", margin = "documents")
 #' textstat_simil2(pres_dfm, "2017-Trump", method = "cosine", margin = "documents")
 #' textstat_simil2(pres_dfm, c("2009-Obama" , "2013-Obama"), margin = "documents")
-#' 
+#'
 #' # compute some term similarities
-#' s2 <- textstat_simil2(pres_dfm, c("fair", "health", "terror"), method = "cosine", 
+#' s2 <- textstat_simil2(pres_dfm, c("fair", "health", "terror"), method = "cosine",
 #'                       margin = "features")
 #' head(as.matrix(s2), 10)
 #' as.list(s2, n = 8)
