### 2018-10-02 22d8adb
Trim features to speed up CI test
