### 2018-10-02 #774 @arfon
@lmullen, @borishejblum - could you both confirm that the changes @kbenoit has made in https://github.com/quanteda/quanteda/pull/1431 have addressed your feedback?
