### 2018-10-02 #774 @borishejblum
@arfon @kbenoit it seems that my comment regarding the definition of who the target audience is has not been addressed (or did I miss something ?).
Thanks
