### 2018-10-03 0f74e48
Reimplement euclidean, hamming, kullback, manhattan, maximum, canberra and minkowski distannt masures.
