+++ tests/testthat/test-textstat_simil2.R
@@ -4,247 +4,200 @@ test_mt <- dfm(corpus_subset(data_corpus_inaugural, Year > 1980), remove = stopw
                stem = TRUE, verbose = FALSE)
 test_mt <- dfm_trim(test_mt, min_termfreq = 5)
 
-# correlation
-test_that("test textstat_simil2 method = correlation on documents", {
-    skip_if_not_installed("proxy")
-    s1 <- as.matrix(textstat_simil2(test_mt, method = "correlation", margin = "documents"))
-    s2 <- as.matrix(proxy::simil(as.matrix(test_mt), by_rows = TRUE, diag = TRUE))
+test_simil <- function(x, method, margin, ignore_upper = FALSE, ...) {
+    
+    if (margin == "documents") {
+        by_rows <- TRUE
+        selection <- "1985-Reagan"
+        y <- x[selection,]
+    } else {
+        by_rows <- FALSE
+        selection <- "soviet"
+        y <- x[,selection]
+    }
+    
+    s1 <- as.matrix(textstat_simil2(x, method = method, margin = margin, ...))
+    s2 <- as.matrix(proxy::simil(as.matrix(x), 
+                                method = method, by_rows = by_rows, diag = TRUE, ...))
+
+    if (ignore_upper)
+        s1[upper.tri(s1, TRUE)] <- s2[upper.tri(s2, TRUE)] <- 0
     expect_equal(s1, s2, tolerance = 0.001)
     
-    s3 <- as.matrix(textstat_simil2(test_mt, "1985-Reagan", method = "correlation", margin = "documents"))
-    s4 <- as.matrix(proxy::simil(as.matrix(test_mt), as.matrix(test_mt["1985-Reagan",]), by_rows = TRUE, diag = TRUE))
+    s3 <- as.matrix(textstat_simil2(x, selection, method = method, margin = margin, ...))
+    s4 <- as.matrix(proxy::simil(as.matrix(x), as.matrix(y), 
+                                method = method, by_rows = by_rows, diag = TRUE, ...))
+    if (ignore_upper)
+        s3[upper.tri(s3, TRUE)] <- s4[upper.tri(s4, TRUE)] <- 0
     expect_equal(as.numeric(s3), as.numeric(s4), tolerance = 0.001)
+}
+
+test_dist <- function(x, method, margin, ignore_upper = FALSE, ...) {
+    
+    if (margin == "documents") {
+        by_rows <- TRUE
+        selection <- "1985-Reagan"
+        y <- x[selection,]
+    } else {
+        by_rows <- FALSE
+        selection <- "soviet"
+        y <- x[,selection]
+    }
+    
+    s1 <- as.matrix(textstat_simil2(x, method = method, margin = margin, ...))
+    s2 <- as.matrix(proxy::dist(as.matrix(x), 
+                                method = method, by_rows = by_rows, diag = TRUE, ...))
+    
+    if (ignore_upper)
+        s1[upper.tri(s1, TRUE)] <- s2[upper.tri(s2, TRUE)] <- 0
+    expect_equal(s1, s2, tolerance = 0.001)
+    
+    s3 <- as.matrix(textstat_simil2(x, selection, method = method, margin = margin, ...))
+    s4 <- as.matrix(proxy::dist(as.matrix(x), as.matrix(y), 
+                                method = method, by_rows = by_rows, diag = TRUE, ...))
+    if (ignore_upper)
+        s3[upper.tri(s3, TRUE)] <- s4[upper.tri(s4, TRUE)] <- 0
+    expect_equal(as.numeric(s3), as.numeric(s4), tolerance = 0.001)
+}
+
+# Similarity measures -------------------------------------------
+
+test_that("test textstat_simil2 cosine similarity", {
+    skip_if_not_installed("proxy")
[…]
