### 2018-10-03 78d6ec9
Add statement of need and target audience
- As requested by @borishejblum in #1393.
- Adds a paragraph about this to both the introduction to the paper and to the `README.Rmd`.
