+++ index.Rmd
@@ -15,6 +15,17 @@ knitr::opts_chunk$set(
 
 **quanteda** is an R package for managing and analyzing textual data developed by [Kenneth Benoit](http://kenbenoit.net) and other contributors. Its initial development was supported by the European Research Council grant ERC-2011-StG 283794-QUANTESS.
 
+The package is designed for R users needing to apply natural language
+processing to texts, from documents to final analysis.  Its
+capabilities match or exceed those provided in many end-user software
+applications, many of which are expensive and not open source.  The package is
+therefore of great benefit to researchers, students, and other analysts with
+fewer financial resources.  While using **quanteda** requires R programming
+knowledge, its API is designed to enable powerful, efficient analysis with a
+minimum of steps.  By emphasising consistent design, furthermore, **quanteda**
+lowers the barriers to learning and using NLP and quantitative text analysis
+even for proficient R programmers.
+
 ## How to Install
 
 The normal way from CRAN, using your R GUI or 
