### 2018-10-03 #774 @kbenoit
@borishejblum please see https://github.com/quanteda/quanteda/pull/1431/commits/78d6ec9e7f000bc20ea27ba8cdad3e814bacff5a
