### 2018-10-03 #774 @borishejblum
Thanks @kbenoit. All my comments have now been succesfully adressed by quanteda's authors.
