### 2018-10-03 afe77f7
Minor adjustment to American English
