### 2018-10-04 7c873c8
Implement chisquare for average-profile weighted matrix for #1442
