### 2018-10-04 a3b5355
Increment version, add news for #1442 fix
