### 2018-10-04 d7ea8c2
Minor fixes for CRAN 1.3.10 submission
- updated cran-comments.md
- fixed URLs that became broken because of target changes to Jurafsky and Martin's online book
- updated NEWS
- ranrevdep check
