+++ R/fcm.R
@@ -73,10 +73,9 @@
 #'   2010 Annual Conference of the North American Chapter of the ACL}, Los 
 #'   Angeles, California, June 2010, pp. 325-328.
 #'   
-#'   Daniel Jurafsky & James H. Martin. (2015) \emph{Speech and Language 
-#'   Processing}.  Draft of April 11, 2016. 
-#'   \href{https://web.stanford.edu/~jurafsky/slp3/16.pdf}{Chapter 16, Semantics
-#'   with Dense Vectors.}
+#'   Jurafsky, Daniel and James H. Martin (2018). "Chapter 6, Vector Semantics."
+#'   from \emph{Speech and Language Processing}.  Draft of September 23, 2018,
+#'   from \url{https://web.stanford.edu/~jurafsky/slp3/}}
 #'   
 #'   Church, K. W. & P. Hanks (1990) 
 #'   "\href{http://dl.acm.org/citation.cfm?id=89095}{Word association norms,
