### 2018-10-05 #774 @lmullen
I've revised the diffs of the changes and re-run tests, R CMD check, and so on. The improvements to the paper have been made, and the improvements to the local tests in particular make a huge difference for checking the package. I'm satisfied that the changes have been made and the paper can be accepted.

Very much looking forward to citing this paper as I continue to use quanteda in my own work. Congratulations to @kbenoit and team on such fantastic software.
