### 2018-10-05 #774 @kbenoit
Thanks @arfon (and once again, to all the reviewers for your excellent and supportive comments.

I've updated the release to v1.3.10 (also under CRAN review) and Zenodo has updated the DOI to 10.5281/zenodo.1447219.
